% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_helpers.R
\name{add_missing_columns}
\alias{add_missing_columns}
\title{Add missing columns}
\usage{
add_missing_columns(data, columns)
}
\arguments{
\item{data}{A data frame.}

\item{columns}{Expected columns in the data frame.}
}
\value{
A data frame with at least all the columns named in \code{columns}.
The added columns will be of class \code{character}.
}
\description{
Small helper function. Checks if all expected columns are in the data frame
and if not, adds a column with all missing variables in the data. Thus, the
columns will be made explicitly missing in the data frame.
}
\examples{
\dontrun{
add_missing_columns(head(iris), c("important_column1", "important_column2"))
}
}
\keyword{internal}
