% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_appdata.R
\name{add_text_label}
\alias{add_text_label}
\title{Add text label}
\usage{
add_text_label(
  data,
  label_name = "text_label",
  value = "item_value",
  item_unit = "item_unit",
  lower_lim = "lower_lim",
  upper_lim = "upper_lim",
  subject_id = "subject_id",
  event_date = "event_date",
  event_name = "event_name",
  day = "day",
  significance = "significance"
)
}
\arguments{
\item{data}{}
}
\description{
Helper function for \code{\link[=get_appdata]{get_appdata()}}.
}
\keyword{internal}
