% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_helpers.R
\name{add_timevars_to_data}
\alias{add_timevars_to_data}
\title{Add time vars to raw data}
\usage{
add_timevars_to_data(data, required_col_names = required_col_names)
}
\arguments{
\item{data}{A data frame}

\item{required_col_names}{Required column names.}
}
\value{
A data frame, with derivative time and event variables, needed for
ClinSight to function properly.
}
\description{
Add time vars to raw data
}
