% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_helpers.R
\name{adjust_colnames}
\alias{adjust_colnames}
\title{Adjust column names}
\usage{
adjust_colnames(data, pattern, replacement = "")
}
\arguments{
\item{data}{A data frame}

\item{pattern}{Pattern to search for in the names.}

\item{replacement}{Replacement for the pattern. Default to empty (thus,
removing the pattern from the data frame names).}
}
\value{
A data frame with adjusted column names.
}
\description{
Small helper function to adjust column names in a data frame. Can be used in
a \code{magrittr} or native \code{R} pipe.
}
\details{
Used in the application in the \code{\link[=create_table]{create_table()}} form-specific S3 methods
to remove common prefixes for better table layout. For example, the prefix
"AE" in front of all column names of the Adverse event table can be removed
with this function.
}
\examples{
adjust_colnames(head(iris), "^Sepal", "Flower")
}
\keyword{internal}
