% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_server.R
\name{app_server}
\alias{app_server}
\title{App Server}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for \code{shiny}.}
}
\description{
Main server of the Shiny application.
}
\details{
Prepares the app data and connects all the Shiny modules. Main modules used
in the application are the \code{common_form} pages (\code{\link[=mod_common_forms_server]{mod_common_forms_server()}}),
\verb{study form} pages (\code{\link[=mod_study_forms_server]{mod_study_forms_server()}}), the \code{sidebar}
(\code{\link[=mod_main_sidebar_server]{mod_main_sidebar_server()}}), the \verb{start page} (\code{\link[=mod_start_page_server]{mod_start_page_server()}}),
the \verb{header widgets} (\code{\link[=mod_header_widgets_server]{mod_header_widgets_server()}}), and the \verb{query page}
(\code{\link[=mod_queries_server]{mod_queries_server()}})
}
\seealso{
\code{\link[=app_ui]{app_ui()}}, \code{\link[=run_app]{run_app()}}
}
\keyword{internal}
