% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_appdata.R
\name{apply_custom_functions}
\alias{apply_custom_functions}
\title{Apply custom modification functions}
\usage{
apply_custom_functions(data, functions = NULL, .default = "identity")
}
\arguments{
\item{data}{A data frame (for example, raw data merged).}

\item{functions}{A character vector containing the names of the functions to
apply to the data. Default is NULL.}

\item{.default}{A character vector containing the names of the functions to
apply if none are provided. Default is "identity".}
}
\description{
Apply custom modification functions
}
\keyword{internal}
