% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_appdata.R
\name{apply_edc_specific_changes}
\alias{apply_edc_specific_changes}
\title{Apply EDC-specific suffix fixes}
\usage{
apply_edc_specific_changes(
  data,
  expected_columns = c("LBORNR_Lower", "LBORNR_Upper", "LBORRESU", "LBORRESUOTH",
    "LBCLSIG", "LBREASND")
)
}
\arguments{
\item{data}{A data frame}

\item{expected_columns}{A character vector with the columns that should be
expected in the data frame. If missing, these columns will be added with
missing data (thus, will be made explicitly missing).}
}
\value{
A data frame.
}
\description{
These changes are study/EDC-specific and part of the legacy code for ClinSight.
}
\keyword{internal}
