% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_appdata.R
\name{apply_study_specific_fixes}
\alias{apply_study_specific_fixes}
\title{Apply study-specific fixes}
\usage{
apply_study_specific_fixes(
  data,
  form_id_vars = c("subject_id", "event_name", "item_group")
)
}
\arguments{
\item{data}{A data frame (for example, raw data merged).}

\item{form_id_vars}{A character vector with the names of the columns that
identify a form.}
}
\value{
A data frame.
}
\description{
These changes are probably study-specific and need to be changed accordingly.
}
