% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_appdata.R
\name{apply_study_specific_suffix_fixes}
\alias{apply_study_specific_suffix_fixes}
\title{Apply study-specific suffix fixes}
\usage{
apply_study_specific_suffix_fixes(data)
}
\arguments{
\item{data}{A data frame}
}
\value{
A data frame.
}
\description{
These changes are study/EDC-specific and part of the legacy code for ClinSight.
}
\keyword{internal}
