% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_authentication.R
\name{authenticate_server}
\alias{authenticate_server}
\title{Authenticate server}
\usage{
authenticate_server(
  user_identification = get_golem_config("user_identification"),
  all_sites = NULL,
  all_roles = get_golem_config("group_roles"),
  credentials_db = file.path(get_golem_config("data_folder"),
    get_golem_config("credentials_db")),
  credentials_pwd = Sys.getenv("DB_SECRET"),
  user_id = get_golem_config("user_id"),
  user_name = get_golem_config("user_name"),
  user_group = get_golem_config("user_group"),
  session
)
}
\arguments{
\item{user_identification}{Character vector showing the user identification.
Is by default set by the \code{user_identification} option in the \code{golem-config}
file.}

\item{all_sites}{Character vector with all sites. Will be passed on to
shinymanager configuration so that data can be restricted to specific sites
per user.}

\item{all_roles}{Character vector with all roles. Used to show all applicable
roles in \code{shinymanager} admin mode.}

\item{credentials_db}{Character vector. Path to the credentials database. By
default, set by the \code{data_folder} and \code{credentials_db} options in the
\code{golem-config_file}.}

\item{credentials_pwd}{Character vector, containing the database password.}

\item{user_id}{Character vector. Used to retrieve the user ID from the
session object, if applicable.}

\item{user_name}{Character vector. Used to retrieve the user name from the
session object, if applicable.}

\item{user_group}{Used to retrieve the user group from the session object, if
applicable.}

\item{session}{Shiny session. Needed to access user information in case of
login methods alternative to \code{shinymanager} are used.}
}
\description{
Function to authenticate the main server.
}
