% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_authentication.R
\name{authenticate_ui}
\alias{authenticate_ui}
\title{Authenticate UI}
\usage{
authenticate_ui(test_mode = FALSE)
}
\arguments{
\item{test_mode}{Logical. Whether the app should be started in test mode or
not.}
}
\description{
Authentication implementation in the UI.
}
