% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{bind_rows_custom}
\alias{bind_rows_custom}
\title{Custom row bind}
\usage{
bind_rows_custom(data, convert = "item_value")
}
\arguments{
\item{data}{A list of data frames.}

\item{convert}{A character vector containing the column names of the columns
that need to be converted.}
}
\value{
A data frame object.
}
\description{
Small wrapper around the function \code{dplyr::bind_rows()}.
Converts first a custom amount of variables in a list of data frames to
character, and then combines the list of data frames into one data frame.
}
\examples{
\dontrun{
 data_list <- list(
   data.frame(
      "ID" = c("subj_1", "subj_5", "subj_8"),
      "values" = c("1", "4", "7")
   ),
   data.frame(
      "ID" = c("subj_7", "subj_9", NA_character_),
      "values" = c(10, 12, 5)
   )
 )
bind_rows_custom(data_list, "values")
}
}
\keyword{internal}
