% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_tests.R
\name{check_appdata}
\alias{check_appdata}
\title{Check data}
\usage{
check_appdata(
  data = appdata,
  meta = metadata,
  required_cols = c("site_code", "subject_id", "event_date", "event_name", "item_group",
    "form_repeat", "item_name", "edit_date_time", "day")
)
}
\arguments{
\item{data}{List of data frames with app data.}

\item{meta}{data frame with meta data.}

\item{required_cols}{Character vector, containing the column names that are
required for the application to run appropriately.}
}
\description{
Checks data for compatibility with metadata.
}
