% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_tests.R
\name{check_available_data}
\alias{check_available_data}
\title{Check available data}
\usage{
check_available_data(
  required_data = c("appdata", "apptables", "vars", "db_path"),
  env = globalenv()
)
}
\arguments{
\item{required_data}{A character vector with objects that are required.}

\item{env}{The environment to check in.}
}
\value{
Nothing if the checks are passed. If not, will throw an error.
}
\description{
Checks if data is loaded in memory and if not, provides an informative error
message.
}
\examples{
# check_available_data("missing_data_frame")
# Error in check_available_data("missing_data_frame") : 
# The following data objects are required to run the app properly, 
# but are missing: missing_data_frame

}
