% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{check_query_path}
\alias{check_query_path}
\title{Check Query Path}
\usage{
check_query_path(query_path)
}
\arguments{
\item{query_path}{If it exists, provide a file.path to an RDS file,
containing data.frame with the same schema as \code{query_data_skeleton}, used
to populate a new database with existing queries from the EDC.}
}
\value{
a data.frame containing EDC queries
}
\description{
Verify that the query_path supplied is in the correct format, and if it
isn't, fall back on the internal query_data_skeleton. If it is, then read
the rds file in as a data.frame
}
\keyword{internal}
