% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{clean_dates}
\alias{clean_dates}
\title{Clean dates}
\usage{
clean_dates(x, unknown_pattern = "NK", unknown_replacement = "01")
}
\arguments{
\item{x}{A character or date vector. if the input is already a date vector,
the input will be returned unchanged.}

\item{unknown_pattern}{Pattern to look up in the date, standard "NK". Note
that the pattern will gain a prefix of "-", to avoid that years with the
tag \code{unknown_pattern} tab will be replaced. Dates containing years with the
\code{unknown_pattern} tag will become missing.}

\item{unknown_replacement}{Replacement of the unknown pattern. Standard 01
(first day/month).}
}
\value{
A \code{date}-class vector.
}
\description{
This helper function replaces a pattern of "NK" in months and in days in a
date, with the first day or month. The function expects that the pattern of
the date is 'YYYY-MM-DD'.
}
\examples{
clean_dates(c("2023-08-NK", "2023-NK-NK", "NK-08-01"))
}
