% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_customize_events.R
\name{clean_event_metadata}
\alias{clean_event_metadata}
\title{Clean event metadata}
\usage{
clean_event_metadata(data, event_id_cols = c("event_id", "event_id_pattern"))
}
\arguments{
\item{data}{A data frame with event data.}

\item{event_id_cols}{A character vector with the event_id columns.}
}
\value{
A data frame with the columns: event_id, event_id_pattern,
event_name_custom, event_label_custom, is_regular_visit, is_baseline_event,
generate_labels, meta_event_order, add_visit_number,
add_event_repeat_number.
}
\description{
Internal function to verify and clean event data.
}
\keyword{internal}
