% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_merge_item_pairs.R
\name{clean_merge_pair_metadata}
\alias{clean_merge_pair_metadata}
\title{Clean merge item pair.}
\usage{
clean_merge_pair_metadata(
  data,
  merge_col = "merge_with",
  suffix_to_add = "_ITEM_TO_MERGE_WITH_PAIR"
)
}
\arguments{
\item{data}{A data frame with metadata}

\item{merge_col}{A character vector with the column name in which the
variable to merge with is defined.}

\item{suffix_to_add}{A Character vector with the suffix to add to the
item_name. This is used to detect item pairs and merge them when in
\code{\link[=merge_meta_with_data]{merge_meta_with_data()}}.}
}
\description{
Called in get_metadata to add additional rows with merge item pairs. This is
needed so that all these 'other' columns are preserved when merging data with
metadata.
}
\keyword{internal}
