% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clinsightful_data}
\alias{clinsightful_data}
\title{Clinical Trial test data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6483 rows and 24 columns.
}
\source{
Created with \code{data-raw/create_random_data.R}
}
\usage{
clinsightful_data
}
\description{
A data frame containing randomly created clinical trial data. Used for
testing purposes.
}
\keyword{datasets}
