% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{col_palette}
\alias{col_palette}
\title{Custom color palette}
\format{
\subsection{\code{col_palette}}{

A named vector with the following components:
\itemize{
\item \verb{within limits} : "#1a9641"
\item \verb{out of limits, clinically insignificant}: "#a6d96a"
\item \verb{out of limits, significance pending}: "#fdae61"
\item \verb{out of limits, clinically significant}: "#d7191c"
\item \verb{limits unknown}: "grey50"
}
}
}
\usage{
col_palette
}
\description{
A named vector of colors to be used in the figures of the application.
}
\keyword{datasets}
