% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_data_fct_helpers.R
\name{collapse_column_vals}
\alias{collapse_column_vals}
\title{Collapse column values}
\usage{
collapse_column_vals(
  data,
  column = "event_name",
  exclude = "",
  group_by = c("subject_id", "item_group"),
  separator = ", "
)
}
\arguments{
\item{data}{A data frame in long format.}

\item{column}{Character string. Column containing the strings to collapse.}

\item{exclude}{Character vector. Values that should not be collapsed.}

\item{group_by}{Character vector. Collapsing will occur grouped by these
variables.}

\item{separator}{Character string to collapse the column values with.}
}
\value{
A data frame in which the \code{column} is changed.
}
\description{
Collapses all the unique values of a character variable in a (long-format)
data frame with a given separator and within a given group. It is also
possible to define values that need to be excluded from the collapse; any
rows with these excluded values will be left as is.
}
\examples{
 library(dplyr)
 df <- data.frame(
  ID = sample(1:5, 15, replace = TRUE),
  name = sample(c("collapsed", "excluded", "names"), 15, replace = TRUE)
  ) |>
  mutate(
    site = sample(c("S1", "S2", "S3"), 1),
    .by = ID
 ) |>
 arrange(ID, factor(name))
 collapse_column_vals(df, column =  "name", exclude = "excluded",
  group_by = c("ID", "site"))
  
}
\keyword{internal}
