% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{collapse_fct_levels}
\alias{collapse_fct_levels}
\title{Collapse Factor levels}
\usage{
collapse_fct_levels(
  x = appdata$ECG$status,
  new_levels = list(old = "old", new = c("new", "updated")),
  warn_non_existent_levels = FALSE
)
}
\arguments{
\item{x}{a character string or factor}

\item{new_levels}{a named list of levels that need to be relabeled. They
should be character values. If no names are provided, the first element
will be chosen as the factor label.}

\item{warn_non_existent_levels}{A logical, whether a warning should be given
if levels are not existent.}
}
\value{
a factor with newly calculated levels
}
\description{
Small helper function to collapse factors using a Base R syntax
}
\examples{
collapse_fct_levels(
x = LETTERS,
new_levels = list("A-C" = c("A", "B", "C"))
)
}
