% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_header_widgets_fct_helpers.R
\name{count_adverse_events}
\alias{count_adverse_events}
\title{Count Adverse Events}
\usage{
count_adverse_events(
  data,
  all_ids = NULL,
  SAE_column_name = "Serious Adverse Event"
)
}
\arguments{
\item{data}{A data frame with Adverse Event data}
}
\value{
A data frame with the columns \code{subject_id}, \code{AEs} (number of AEs per
subject), \code{SAEs} (number of SAEs per subject).
}
\description{
Simple helper function to count Adverse Events (AEs) and Serious Adverse
Events (SAEs).
}
\keyword{internal}
