% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_customize_clinsight_helpers.R
\name{create_clinsight_config}
\alias{create_clinsight_config}
\title{Create custom ClinSight configuration file}
\usage{
create_clinsight_config(
  path = ".",
  template_path = app_sys("golem-config.yml")
)
}
\arguments{
\item{path}{Character string with destination path.}

\item{template_path}{Path to the config template.}
}
\value{
A .yml configuration file.
}
\description{
Small helper function to create a custom ClinSight configuration file in a
temporary folder. It will attempt to open it.
}
