% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_customize_clinsight_helpers.R
\name{create_clinsight_metadata}
\alias{create_clinsight_metadata}
\title{Create custom ClinSight metadata}
\usage{
create_clinsight_metadata(path = ".", template_path = app_sys("metadata.xlsx"))
}
\arguments{
\item{path}{Character string with destination path.}

\item{template_path}{Path to the metadata template.}
}
\value{
A ClinSight metadata (.xlsx) file.
}
\description{
Small helper function to open a new metadata Excel file in a temporary
folder. It will attempt to open it.
}
