% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{create_delete_log_trigger}
\alias{create_delete_log_trigger}
\title{Create Delete Logging Trigger}
\usage{
create_delete_log_trigger(con)
}
\arguments{
\item{con}{A DBI Connection to the SQLite DB}
}
\description{
Logs record deletions on all_review_data into all_review_data_log.
}
\keyword{internal}
