% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report_fct_helpers.R
\name{create_report}
\alias{create_report}
\title{Create Report}
\usage{
create_report(fileinput, reviewer, study_sites, review_df, query_df)
}
\arguments{
\item{fileinput}{Character vector with the name of the file to be created.}

\item{reviewer}{Character vector with the name of the current reviewer.}

\item{study_sites}{Character vector with the names of the sites that were reviewed.}

\item{review_df}{Data frame with review data to be used in the report.}

\item{query_df}{Data frame with query data to be used in the report.}
}
\value{
A PDF report will be created.
}
\description{
Creates a PDF markdown report.
}
\keyword{internal}
