% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tables.R
\name{create_table.adverse_events}
\alias{create_table.adverse_events}
\title{Create Adverse Events table}
\usage{
\method{create_table}{adverse_events}(
  data,
  name_column = "item_name",
  value_column = "item_value",
  keep_vars = c("subject_id", "form_repeat"),
  expected_columns = NULL,
  worsening_start_column = "date of worsening",
  SAE_start_column = "SAE Start date",
  AE_start_column = "start date",
  AE_end_column = "end date",
  ...
)
}
\arguments{
\item{data}{A data frame to convert, in long format.}

\item{name_column}{Character string. Column that contains all the names for
the wide table format.}

\item{value_column}{Character string. Column that contains all the values of
the table in wide format.}

\item{keep_vars}{all common variables that identify a unique single data row.}

\item{expected_columns}{A character vector with the names of the expected
variables. Needed to make sure that all expected columns are always
created, even if some variables are implicitly missing (which might occur
if there are not yet any values available for a specific variable). Also,
implicitly missing variables might give errors if part of the script relies
on the variables' presence.}

\item{worsening_start_column}{Character vector. Name of the column containing
the date at which worsening of the (S)AE started.}

\item{SAE_start_column}{Character vector. Name of the column containing the
SAE start date.}

\item{AE_start_column}{Character vector. Name of the column containing the AE
start date.}

\item{AE_end_column}{Character vector. Name of the column containing the AE
end date.}

\item{...}{Other options. Currently unused.}
}
\value{
A data frame in wide format
}
\description{
Function to create an adverse event dataset.
}
