% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tables.R
\name{create_table.common_forms}
\alias{create_table.common_forms}
\title{Create Default 'Common Events' Table}
\usage{
\method{create_table}{common_forms}(
  data,
  name_column = "item_name",
  value_column = "item_value",
  keep_vars = c("subject_id", "form_repeat"),
  expected_columns = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame to convert, in long format.}

\item{name_column}{Character string. Column that contains all the names for
the wide table format.}

\item{value_column}{Character string. Column that contains all the values of
the table in wide format.}

\item{keep_vars}{all common variables that identify a unique single data row.}

\item{expected_columns}{A character vector with the names of the expected
variables. Needed to make sure that all expected columns are always
created, even if some variables are implicitly missing (which might occur
if there are not yet any values available for a specific variable). Also,
implicitly missing variables might give errors if part of the script relies
on the variables' presence.}

\item{...}{Other options. Currently unused.}
}
\value{
A data frame in wide format
}
\description{
Create Default 'Common Events' Table
}
