% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tables.R
\name{create_table.continuous}
\alias{create_table.continuous}
\title{Create Table with continuous data.}
\usage{
\method{create_table}{continuous}(
  data,
  name_column = "item_name",
  value_column = "item_value",
  unit_column = "item_unit",
  explanation_column = "reason_notdone",
  keep_vars = c("subject_id", "event_name", "event_date"),
  expected_columns = NULL,
  ...
)
}
\arguments{
\item{data}{A data frame to convert, in long format.}

\item{name_column}{Character string. Column that contains all the names for
the wide table format.}

\item{value_column}{Character string. Column that contains all the values of
the table in wide format.}

\item{unit_column}{Character vector. The name of the column that contains
units (as character values).}

\item{explanation_column}{Character vector. The name of the column that
contains an explanation why a value is missing (if available).}

\item{keep_vars}{all common variables that identify a unique single data row.}

\item{expected_columns}{A character vector with the names of the expected
variables. Needed to make sure that all expected columns are always
created, even if some variables are implicitly missing (which might occur
if there are not yet any values available for a specific variable). Also,
implicitly missing variables might give errors if part of the script relies
on the variables' presence.}

\item{...}{Other options. Currently unused.}
}
\value{
A data frame in wide format
}
\description{
Create Table with continuous data.
}
