% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_tables.R
\name{create_table}
\alias{create_table}
\title{Create table}
\usage{
create_table(data, ...)
}
\arguments{
\item{data}{A data frame to be converted.
Based on the custom class assigned to the table, a different
transformation will be used.}

\item{...}{Other possible options}
}
\value{
A table in wide format.
}
\description{
Creates a wide form data table based on custom S3 classes.
}
\seealso{
\code{\link[=create_table.default]{create_table.default()}}, \code{\link[=create_table.continuous]{create_table.continuous()}},
\code{\link[=create_table.adverse_events]{create_table.adverse_events()}}, \code{\link[=create_table.medication]{create_table.medication()}},
\code{\link[=create_table.medical_history]{create_table.medical_history()}}, \code{\link[=create_table.bm_cytology]{create_table.bm_cytology()}},
\code{\link[=create_table.conc_procedures]{create_table.conc_procedures()}}.
}
