% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{create_unique_id}
\alias{create_unique_id}
\title{Create Unique ID}
\usage{
create_unique_id(bytes = 5)
}
\arguments{
\item{bytes}{An integer. Controls the size of the unique id.}
}
\value{
A unique character vector.
}
\description{
Creates a unique Identifier. Adapted from \code{crosstalk:::createUniqueId()} since
it is not an exported function. This id needs to be refreshed at certain times
to ensure that the crosstalk widgets stay connected (see also
\href{https://github.com/rstudio/crosstalk/issues/132#issue-1393291183}{here}).
}
\examples{
create_unique_id(5)
}
\keyword{internal}
