% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs_app_tables}
\alias{cs_app_tables}
\title{Application Tables}
\format{
\subsection{\code{cs_app_tables}: A list with 11 objects.}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{str(cs_app_tables)
#> List of 11
#>  $ Adverse events  : tibble [90 x 16] (S3: tbl_df/tbl/data.frame)
#>   ..$ subject_id           : chr [1:90] "BEL_07_193" "NLD_03_207" "BEL_08_885" "NLD_03_207" ...
#>   ..$ form_repeat          : int [1:90] 1 1 1 1 1 2 2 2 3 6 ...
#>   ..$ Name                 : chr [1:90] "Atelectasis" "Sepsis" "Seizure" "Sepsis" ...
#>   ..$ AESI                 : chr [1:90] "None" "None" "None" "None" ...
#>   ..$ start date           : chr [1:90] "2023-09-14" "2023-08-14" "2023-08-11" "2023-08-02" ...
#>   ..$ end date             : chr [1:90] NA "2023-08-14" NA "2023-08-14" ...
#>   ..$ CTCAE severity       : chr [1:90] "Grade 3" "Grade 5" "Grade 3" "Grade 3" ...
#>   ..$ Treatment related    : chr [1:90] "Unlikely" "Not related" "Not related" "Not related" ...
#>   ..$ Treatment action     : chr [1:90] "Not Applicable (Concomitant therapy/medication; Hospitalization/Prolongation of hospitalization)" "Interrupted (Concomitant therapy/medication; Hospitalization/Prolongation of hospitalization)" "Not Applicable (Hospitalization/Prolongation of hospitalization)" "Interrupted (Concomitant therapy/medication; Hospitalization/Prolongation of hospitalization)" ...
#>   ..$ Serious Adverse Event: chr [1:90] "Yes" "Yes" "Yes" "Yes" ...
#>   ..$ SAE Start date       : chr [1:90] "2023-09-14" "2023-08-14" "2023-08-11" "2023-08-02" ...
#>   ..$ SAE End date         : chr [1:90] NA "2023-08-14" NA "2023-08-14" ...
#>   ..$ SAE Category         : chr [1:90] "Requires inpatient hospitalization or prolongation of existing hospitalization" "Results in death; Is life-threatening; Requires inpatient hospitalization or prolongation of existing hospitalization" "Requires inpatient hospitalization or prolongation of existing hospitalization" "Results in death; Is life-threatening; Requires inpatient hospitalization or prolongation of existing hospitalization" ...
#>   ..$ SAE Awareness date   : chr [1:90] "2023-09-14" "2023-08-02" "2023-08-15" "2023-08-02" ...
#>   ..$ SAE Date of death    : chr [1:90] NA "2023-08-14" NA "2023-08-14" ...
#>   ..$ SAE Death reason     : chr [1:90] NA "disease progression" NA "disease progression" ...
#>  $ CBC regular     : tibble [65 x 9] (S3: tbl_df/tbl/data.frame)
#>   ..$ subject_id : chr [1:65] "BEL_04_133" "BEL_04_133" "BEL_04_133" "BEL_04_133" ...
#>   ..$ event_name : chr [1:65] "Screening" "Visit 1" "Visit 2" "Visit 3" ...
#>   ..$ event_date : Date[1:65], format: "2023-06-05" "2023-07-05" ...
#>   ..$ Haemoglobin: chr [1:65] "11.5 g/dL" "8 g/dL" "8.2 g/dL" "10.7 g/dL" ...
#>   ..$ Haematocrit: chr [1:65] "0.25 Proportion of 1.0" "0.23 Proportion of 1.0" "0.33 Proportion of 1.0" "0.21 Proportion of 1.0" ...
#>   ..$ Platelets  : chr [1:65] "99.14 10^9/L" "99.14 10^9/L" "99.14 10^9/L" "99.14 10^9/L" ...
#>   ..$ Neutrophils: chr [1:65] "4.16 10^9/L" "1.36 10^9/L" "1.36 10^9/L" "1.36 10^9/L" ...
#>   ..$ Lymphocytes: chr [1:65] "1.7 10^9/L" "1.55 10^9/L" "1.6 10^9/L" "1.09 10^9/L" ...
#>   ..$ Monocytes  : chr [1:65] "0.12 10^9/L" "0.1 10^9/L" "0.83 10^9/L" "0.12 10^9/L" ...
#>  $ Conc. Procedures: tibble [4 x 9] (S3: tbl_df/tbl/data.frame)
#>   ..$ subject_id             : chr [1:4] "BEL_07_645" "BEL_07_645" "DEU_02_866" "NLD_06_755"
#>   ..$ form_repeat            : int [1:4] 1 2 1 1
#>   ..$ Name                   : chr [1:4] "Appendectomy" "Amputation" "Thrombolysis" "Amputation"
#>   ..$ Indication             : chr [1:4] "Prophylaxis" "Pain" "Hypertension" "Infection"
#>   ..$ Start Date             : chr [1:4] "2019-NK-NK" "2019-NK-NK" "2023-08-18" "2023-08-04"
#>   ..$ Ongoing                : chr [1:4] "No" "No" "No" "Yes"
#>   ..$ End Date               : chr [1:4] "2019-NK-NK" "2019-NK-NK" "2023-08-18" NA
#>   ..$ Related Medical History: chr [1:4] "Sickle cell anemia" "Sickle cell anemia" "Epilepsy" "Familial Mediterranean fever"
#>   ..$ Related AE             : chr [1:4] NA NA NA "Joint pain"
#>  $ Electrolytes    : tibble [46 x 9] (S3: tbl_df/tbl/data.frame)
#>   ..$ subject_id : chr [1:46] "BEL_04_133" "BEL_04_133" "BEL_04_133" "BEL_04_133" ...
#>   ..$ event_name : chr [1:46] "Screening" "Visit 1" "Visit 3" "Visit 5" ...
#>   ..$ event_date : Date[1:46], format: "2023-06-05" "2023-07-05" ...
#>   ..$ Sodium     : chr [1:46] "138 mmol/L" "138 mmol/L" "140 mmol/L" "138 mmol/L" ...
#>   ..$ Potassium  : chr [1:46] "4.5 mmol/L" "4.14 mmol/L" "4.65 mmol/L" "4.74 mmol/L" ...
#>   ..$ Chloride   : chr [1:46] "105 mmol/L" "103 mmol/L" "105 mmol/L" "104 mmol/L" ...
#>   ..$ Bicarbonate: chr [1:46] "20.5 mmol/L" "23 mmol/L" "23.7 mmol/L" "21.4 mmol/L" ...
#>   ..$ Calcium    : chr [1:46] "2.26 mmol/L" "2.53 mmol/L" "2.43 mmol/L" "2.39 mmol/L" ...
#>   ..$ Magnesium  : chr [1:46] "0.87 mmol/L" "0.83 mmol/L" "0.75 mmol/L" "0.88 mmol/L" ...
#>  $ General         : tibble [25 x 23] (S3: tbl_df/tbl/data.frame)
#>   ..$ subject_id             : chr [1:25] "BEL_04_133" "BEL_04_772" "BEL_07_193" "BEL_07_431" ...
#>   ..$ Age                    : chr [1:25] "88" "78" "26" "42" ...
#>   ..$ Sex                    : chr [1:25] "Male" "Male" "Female" "Male" ...
#>   ..$ ECOG                   : chr [1:25] "1" "0" "1" "1" ...
#>   ..$ Eligible               : chr [1:25] "Yes" "Yes" "Yes" NA ...
#>   ..$ Eligible_Date          : chr [1:25] "2023-07-06" "2023-08-17" "2023-08-23" NA ...
#>   ..$ WHO.classification     : chr [1:25] "Syndrome K" "Syndrome O" "Syndrome D" NA ...
#>   ..$ WHO.subclassification  : chr [1:25] NA NA NA NA ...
#>   ..$ Race                   : chr [1:25] "White" "White" "Not reported" "Not reported" ...
#>   ..$ ChildbearingPotential  : chr [1:25] "No" "No" NA NA ...
#>   ..$ MenopauseReason        : chr [1:25] "Postmenopausal" "Postmenopausal" NA NA ...
#>   ..$ DiscontinuationDate    : chr [1:25] NA NA NA NA ...
#>   ..$ DiscontinuationReason  : chr [1:25] NA NA NA NA ...
#>   ..$ DisconDeathDate        : chr [1:25] NA NA NA NA ...
#>   ..$ DrugAdminDate          : chr [1:25] NA NA NA NA ...
#>   ..$ DrugAdminDose          : chr [1:25] NA NA NA NA ...
#>   ..$ DoseModificationDate   : chr [1:25] NA NA NA NA ...
#>   ..$ DoseModificationReason : chr [1:25] NA NA NA NA ...
#>   ..$ DoseModificationNewDose: chr [1:25] NA NA NA NA ...
#>   ..$ DrugDiscontDate        : chr [1:25] NA NA NA NA ...
#>   ..$ DrugDiscontReason      : chr [1:25] NA NA NA NA ...
#>   ..$ subject_status         : chr [1:25] "Enrolled" "Enrolled" "Enrolled" "Unknown" ...
#>   ..$ status_label           : chr [1:25] "<b>BEL_04_133</b><br><b>Sex:</b> Male<br><b>Age:</b> 88yrs.<br><b>Status:</b> Enrolled<br><b>ECOG:</b> 1<br><b>"| __truncated__ "<b>BEL_04_772</b><br><b>Sex:</b> Male<br><b>Age:</b> 78yrs.<br><b>Status:</b> Enrolled<br><b>ECOG:</b> 0<br><b>"| __truncated__ "<b>BEL_07_193</b><br><b>Sex:</b> Female<br><b>Age:</b> 26yrs.<br><b>Status:</b> Enrolled<br><b>ECOG:</b> 1<br><"| __truncated__ "<b>BEL_07_431</b><br><b>Sex:</b> Male<br><b>Age:</b> 42yrs.<br><b>Status:</b> Unknown<br><b>ECOG:</b> 1<br><b>Dx:</b> NA" ...
#>  $ Liver function  : tibble [45 x 7] (S3: tbl_df/tbl/data.frame)
#>   ..$ subject_id: chr [1:45] "BEL_04_133" "BEL_04_133" "BEL_04_133" "BEL_04_133" ...
#>   ..$ event_name: chr [1:45] "Screening" "Visit 1" "Visit 3" "Visit 5" ...
#>   ..$ event_date: Date[1:45], format: "2023-06-05" "2023-07-05" ...
#>   ..$ Bilirubin : chr [1:45] "6 µmol/L" "11 µmol/L" "8 µmol/L" "7 µmol/L" ...
#>   ..$ AST       : chr [1:45] "21 U/L" "32 U/L" "43 U/L" "42 U/L" ...
#>   ..$ ALT       : chr [1:45] "35 U/L" "45 U/L" "19 U/L" "27 U/L" ...
#>   ..$ gGT       : chr [1:45] "24 U/L" "25 U/L" "26 U/L" "21 U/L" ...
#>  $ Medical History : tibble [152 x 8] (S3: tbl_df/tbl/data.frame)
#>   ..$ subject_id : chr [1:152] "BEL_04_133" "BEL_04_772" "BEL_04_772" "BEL_04_772" ...
#>   ..$ form_repeat: int [1:152] 1 1 2 3 4 5 6 7 8 9 ...
#>   ..$ Name       : chr [1:152] "Epilepsy" "Hypertension" "COPD" "Osteoporosis" ...
#>   ..$ Start Date : chr [1:152] NA "2020-NK-NK" "1995-NK-NK" "2005-NK-NK" ...
#>   ..$ Ongoing    : chr [1:152] "Yes" "Yes" "Yes" "Yes" ...
#>   ..$ End Date   : chr [1:152] NA NA NA NA ...
#>   ..$ Treatment  : chr [1:152] "Medication" "Medication" "Medication" "Medication" ...
#>   ..$ Comment    : chr [1:152] NA NA NA NA ...
#>  $ Medication      : tibble [283 x 10] (S3: tbl_df/tbl/data.frame)
#>   ..$ subject_id             : chr [1:283] "DEU_01_541" "DEU_01_541" "BEL_07_645" "BEL_07_645" ...
#>   ..$ form_repeat            : int [1:283] 1 2 6 7 9 10 3 10 15 33 ...
#>   ..$ Name                   : chr [1:283] "Lisinopril (Sulfaipoetin)" "Oxycodone (Estrizotan)" "Lisinopril (Salsteridone)" "Lisinopril (Nifurasiban)" ...
#>   ..$ Indication             : chr [1:283] "Insomnia" "Infection" "Nausea" "Insomnia" ...
#>   ..$ Dose                   : chr [1:283] "300 Milligram (mg) Per day (QD); Oral (p.o.)" "500 Milligram (mg) Per day (QD); Oral (p.o.)" "5 Milligram (mg) Per month (QM); Intravenous (i.v.)" "20 Milligram (mg) Per month (QM); Intravenous (i.v.)" ...
#>   ..$ Start Date             : chr [1:283] "2023-NK-NK" "2023-NK-NK" "2023-09-13" "2023-09-13" ...
#>   ..$ Ongoing                : chr [1:283] "Yes" "Yes" "Yes" "Yes" ...
#>   ..$ End Date               : chr [1:283] NA NA NA NA ...
#>   ..$ Related Medical History: chr [1:283] NA NA NA NA ...
#>   ..$ Related AE             : chr [1:283] NA NA NA NA ...
#>  $ Renal function  : tibble [46 x 6] (S3: tbl_df/tbl/data.frame)
#>   ..$ subject_id: chr [1:46] "BEL_04_133" "BEL_04_133" "BEL_04_133" "BEL_04_133" ...
#>   ..$ event_name: chr [1:46] "Screening" "Visit 1" "Visit 3" "Visit 5" ...
#>   ..$ event_date: Date[1:46], format: "2023-06-05" "2023-07-05" ...
#>   ..$ Creatinine: chr [1:46] "58 µmol/L" "110 µmol/L" "61 µmol/L" "117 µmol/L" ...
#>   ..$ eGFR      : chr [1:46] "48 ml/min/1.73 m²" "89 ml/min/1.73 m²" "90 ml/min/1.73 m²" "47 ml/min/1.73 m²" ...
#>   ..$ Urate     : chr [1:46] "288 µmol/L" "4.1 mmol/L" "7.5 mmol/L" "294 µmol/L" ...
#>  $ Response        : tibble [3 x 6] (S3: tbl_df/tbl/data.frame)
#>   ..$ subject_id  : chr [1:3] "BEL_04_133" "BEL_04_772" "NLD_06_72"
#>   ..$ event_name  : chr [1:3] "Visit 5" "Visit 5" "Visit 5"
#>   ..$ event_repeat: int [1:3] 5 5 5
#>   ..$ event_date  : Date[1:3], format: "2023-08-30" "2023-08-30" ...
#>   ..$ Response    : chr [1:3] "Stable disease (SD)" "Cured" "Stable disease (SD)"
#>   ..$ Responder   : chr [1:3] "No" "Yes" "No"
#>  $ Vital signs     : tibble [68 x 11] (S3: tbl_df/tbl/data.frame)
#>   ..$ subject_id                   : chr [1:68] "BEL_04_133" "BEL_04_133" "BEL_04_133" "BEL_04_133" ...
#>   ..$ event_name                   : chr [1:68] "Screening" "Visit 1" "Visit 2" "Visit 3" ...
#>   ..$ event_date                   : Date[1:68], format: "2023-06-05" "2023-07-05" ...
#>   ..$ Systolic blood pressure      : chr [1:68] "108 mmHg" "123 mmHg" "122 mmHg" "137 mmHg" ...
#>   ..$ Diastolic blood pressure     : chr [1:68] "62 mmHg" "69 mmHg" "68 mmHg" "72 mmHg" ...
#>   ..$ Pulse                        : chr [1:68] "60 beats/min" "67 beats/min" "69 beats/min" "78 beats/min" ...
#>   ..$ Resp                         : chr [1:68] "19 breaths/min" "19 breaths/min" "18 breaths/min" "16 breaths/min" ...
#>   ..$ Temperature                  : chr [1:68] "36.7 °C" "36.5 °C" "36.2 °C" "36.7 °C" ...
#>   ..$ Weight change since screening: chr [1:68] "0 \%" "24 \%" NA "2 \%" ...
#>   ..$ BMI                          : chr [1:68] "24.52 kg/m2" "17.1 kg/m2" NA "24.52 kg/m2" ...
#>   ..$ Weight                       : chr [1:68] "50 kg" "62 kg" NA "51 kg" ...
}\if{html}{\out{</div>}}
}
}
\source{
See \code{raw-data/clinsightful_data.R} for details.
}
\usage{
cs_app_tables
}
\description{
A list of data frames containing all application tables.
}
\keyword{datasets}
