% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs_app_vars}
\alias{cs_app_vars}
\title{Application Variables}
\format{
\subsection{\code{cs_app_vars}: A list with 6 objects.}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{str(cs_app_vars)
#> List of 6
#>  $ items          :List of 11
#>   ..$ Adverse events  : Named chr [1:17] "ae_number" "ae_name" "aesi" "ae_start_date" ...
#>   .. ..- attr(*, "names")= chr [1:17] "AE Number" "AE Name" "AESI" "AE start date" ...
#>   ..$ CBC regular     : Named chr [1:6] "haemoglobin" "haematocrit" "platelets" "neutrophils" ...
#>   .. ..- attr(*, "names")= chr [1:6] "Haemoglobin" "Haematocrit" "Platelets" "Neutrophils" ...
#>   ..$ Conc. Procedures: Named chr [1:8] "cp_number" "cp_name" "cp_indication" "cp_start_date" ...
#>   .. ..- attr(*, "names")= chr [1:8] "CP Number" "CP Name" "CP Indication" "CP Start Date" ...
#>   ..$ Electrolytes    : Named chr [1:6] "sodium" "potassium" "chloride" "bicarbonate" ...
#>   .. ..- attr(*, "names")= chr [1:6] "Sodium" "Potassium" "Chloride" "Bicarbonate" ...
#>   ..$ General         : Named chr [1:20] "age" "sex" "ecog" "eligible" ...
#>   .. ..- attr(*, "names")= chr [1:20] "Age" "Sex" "ECOG" "Eligible" ...
#>   ..$ Liver function  : Named chr [1:4] "bilirubin" "ast" "alt" "ggt"
#>   .. ..- attr(*, "names")= chr [1:4] "Bilirubin" "AST" "ALT" "gGT"
#>   ..$ Medical History : Named chr [1:7] "mh_number" "mh_name" "mh_start_date" "mh_ongoing" ...
#>   .. ..- attr(*, "names")= chr [1:7] "MH Number" "MH Name" "MH Start Date" "MH Ongoing" ...
#>   ..$ Medication      : Named chr [1:12] "cm_number" "cm_name" "cm_indication" "cm_dose" ...
#>   .. ..- attr(*, "names")= chr [1:12] "CM Number" "CM Name" "CM Indication" "CM Dose" ...
#>   ..$ Renal function  : Named chr [1:3] "creatinine" "egfr" "urate"
#>   .. ..- attr(*, "names")= chr [1:3] "Creatinine" "eGFR" "Urate"
#>   ..$ Response        : Named chr [1:2] "response" "responder"
#>   .. ..- attr(*, "names")= chr [1:2] "Response" "Responder"
#>   ..$ Vital signs     : Named chr [1:8] "systolic_blood_pressure" "diastolic_blood_pressure" "pulse" "resp" ...
#>   .. ..- attr(*, "names")= chr [1:8] "Systolic blood pressure" "Diastolic blood pressure" "Pulse" "Resp" ...
#>  $ all_forms      :'data.frame':	10 obs. of  2 variables:
#>   ..$ main_tab: chr [1:10] "Common events" "Common events" "Common events" "Common events" ...
#>   ..$ form    : chr [1:10] "Adverse events" "Medication" "Conc. Procedures" "Medical History" ...
#>  $ subject_id     : chr [1:25] "BEL_08_45" "BEL_04_133" "BEL_04_772" "BEL_07_193" ...
#>  $ Sites          : advrs_vn [9 x 2] (S3: adverse_events/common_forms/tbl_df/tbl/data.frame)
#>   ..$ site_code: chr [1:9] "BEL04" "BEL07" "BEL08" "BEL09" ...
#>   ..$ region   : chr [1:9] "BEL" "BEL" "BEL" "BEL" ...
#>   ..- attr(*, "synch_time")= chr "2023-09-15 10:10:00 UTC"
#>  $ table_names    : Named chr [1:16] "edit_date_time" "event_date" "event_label" "event_name" ...
#>   ..- attr(*, "names")= chr [1:16] "Edit date" "Date" "Event" "Event" ...
#>  $ form_level_data:'data.frame':	11 obs. of  4 variables:
#>   ..$ item_group         : chr [1:11] "Adverse events" "Medication" "Conc. Procedures" "Medical History" ...
#>   ..$ item_scale         : logi [1:11] NA NA NA NA FALSE TRUE ...
#>   ..$ use_unscaled_limits: logi [1:11] NA NA NA NA TRUE FALSE ...
#>   ..$ review_required    : logi [1:11] TRUE TRUE TRUE TRUE TRUE TRUE ...
}\if{html}{\out{</div>}}
}
}
\source{
See \code{raw-data/clinsightful_data.R} for details.
}
\usage{
cs_app_vars
}
\description{
Named list version of \code{metadata} object.
}
\keyword{datasets}
