% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs_available_data}
\alias{cs_available_data}
\title{Application 'Available Data'}
\format{
a data.frame with 2235 rows and
7 variables.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{str(cs_available_data)
#> 'data.frame':	2235 obs. of  7 variables:
#>  $ subject_id : chr  "BEL_04_772" "BEL_04_772" "BEL_04_772" "BEL_04_772" ...
#>  $ item_name  : chr  "Hypotension" "Atrial Fibrillation (N: 2)" "Tachycardia" "Urinary Tract Infection" ...
#>  $ form_repeat: int  1 2 3 4 5 1 1 1 2 3 ...
#>  $ item_group : chr  "Adverse events" "Adverse events" "Adverse events" "Adverse events" ...
#>  $ event_name : chr  "Any visit" "Any visit" "Any visit" "Any visit" ...
#>  $ event_label: Factor w/ 8 levels "SCR","V1","V2",..: NA NA NA NA NA NA NA NA NA NA ...
#>  $ event_date : Date, format: "2023-08-31" "2023-08-31" ...
#>  - attr(*, "synch_time")= chr "2023-09-15 10:10:00 UTC"
}\if{html}{\out{</div>}}
}
\source{
See \code{raw-data/clinsightful_data.R} for details.
}
\usage{
cs_available_data
}
\description{
Data frame containing info about available data per individual,
such as visits, adverse events, etc. Will be used in module
\code{\link[=mod_queries_server]{mod_queries_server()}}, to select available items to create a query for per
individual and per form. Required columns are the ones distinctively
identifying an item. For now that are site_code, event_name, subject_id,
event_label, item_group, item_name.
}
\keyword{datasets}
