% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs_timeline_data}
\alias{cs_timeline_data}
\title{Application Timeline Data}
\format{
\subsection{\code{cs_timeline_data}: data.frame formatted for consumption by}{

\code{timevis} package.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{str(cs_timeline_data)
#> 'data.frame':	203 obs. of  11 variables:
#>  $ subject_id : chr  "BEL_04_133" "BEL_04_133" "BEL_04_133" "BEL_04_133" ...
#>  $ content    : chr  "Screening" "Visit 1" "Visit 2" "Visit 3" ...
#>  $ form_repeat: int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ item_group : chr  NA NA NA NA ...
#>  $ start      : Date, format: "2023-06-05" "2023-07-05" ...
#>  $ group      : Factor w/ 4 levels "SAE","Adverse event",..: 4 4 4 4 4 4 4 4 4 4 ...
#>  $ end        : Date, format: NA NA ...
#>  $ title      : chr  "2023-06-05 | Screening" "2023-07-05 | Visit 1" "2023-07-19 | Visit 2" "2023-08-02 | Visit 3" ...
#>  $ className  : chr  "bg-light" "bg-light" "bg-light" "bg-light" ...
#>  $ id         : int  1 2 3 4 5 6 7 8 9 10 ...
#>  $ order      : num  4 4 4 4 4 4 4 4 4 4 ...
#>  - attr(*, "synch_time")= chr "2023-09-15 10:10:00 UTC"
}\if{html}{\out{</div>}}
}
}
\source{
See \code{raw-data/clinsightful_data.R} for details.
}
\usage{
cs_timeline_data
}
\description{
A data.frame containing timeline data.
}
\keyword{datasets}
