% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{date_cols_to_char}
\alias{date_cols_to_char}
\title{Convert date columns to character}
\usage{
date_cols_to_char(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame with all columns with the class \code{Date} or \code{POSIXct}
converted to character.
}
\description{
Helpful for when a date object needs to be printed or stored in a SQL
database, since SQL databases do not (always) support date-time classes
used in R.
}
\examples{
\dontrun{
df <- iris
df$dates <- Sys.Date() + sample(1:1000, size = nrow(iris))
converted_df <- date_cols_to_char(df)
# the date variable is now converted, while all other columns are unchanged:
sapply(converted_df, class)
}

}
\keyword{internal}
