% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_add_log}
\alias{db_add_log}
\title{Add Logging Table}
\usage{
db_add_log(con, key_cols = c("id", key_columns))
}
\arguments{
\item{con}{A DBI Connection to the SQLite DB}

\item{key_cols}{An optional character vector specifying which columns should
not be updated in a table. If unset, defaults to 'id' and the
package-defined \code{\link[=key_columns]{key_columns()}}.}
}
\description{
Both creates the logging table and the trigger to update it for
all_review_data.
}
\keyword{internal}
