% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_add_primary_key}
\alias{db_add_primary_key}
\title{Add primary key field}
\usage{
db_add_primary_key(con, name, value, keys = NULL)
}
\arguments{
\item{con}{A DBI Connection to the SQLite DB}

\item{name}{The table name}

\item{value}{A data.frame to add to the table}

\item{keys}{A character vector specifying which columns define a unique row
for the table. If \code{NULL}, no unique index will be created.}
}
\description{
Add primary key field
}
\keyword{internal}
