% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_add_tables}
\alias{db_add_tables}
\title{Add new tables to DB}
\usage{
db_add_tables(con, pk_data, unique_cols, other_data)
}
\arguments{
\item{con}{A DBI Connection to the SQLite DB}

\item{pk_data}{A named list of data frames to add a primary key field to DB
table. Names will correspond to the DB table names.}

\item{unique_cols}{A named list of the fields defining unique records for a
table. Names will correspond to the table to apply the index constraint.}

\item{other_data}{A named list of other data frames to add to the DB. Names
will correspond to the DB table names.}
}
\description{
Add new tables to DB
}
\keyword{internal}
