% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_create}
\alias{db_create}
\title{Create app database}
\usage{
db_create(
  data,
  db_path,
  reviewed = "No",
  reviewer = "",
  status = "new",
  query_path = NULL
)
}
\arguments{
\item{data}{A data frame with review data (Usually created with
\code{\link[=get_review_data]{get_review_data()}}).}

\item{db_path}{A character vector with the path to the database to be
created.}

\item{reviewed}{Character vector. Sets the reviewed tag in the review
database.}

\item{reviewer}{Character vector. Sets the reviewer in the review database.}

\item{status}{Character vector. Sets the status in the review database.
Defaults to \code{new}.}

\item{query_path}{If it exists, provide a file.path to an RDS file,
containing data.frame with the same schema as \code{query_data_skeleton}, used
to populate a new database with existing queries from the EDC.}
}
\value{
A database will be created. Nothing else will be returned.
}
\description{
Creates application database. To create a database with all data flagged as
'new', use the default settings of \code{reviewed}, \code{reviewer}, and \code{status}.
}
\seealso{
\code{\link[=get_review_data]{get_review_data()}}
}
