% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_delete}
\alias{db_delete}
\title{DELETE FROM all_review_data}
\usage{
db_delete(con, data, key_cols = "id")
}
\arguments{
\item{con}{A DBI Connection to the SQLite DB}

\item{data}{A data frame containing the records to DELETE FROM
all_review_data}

\item{key_cols}{A character vectory specifying which columns define a unique
index for a row. Defaults to "id"}
}
\value{
invisible returns TRUE. Is run for its side effects on the DB.
}
\description{
Performs a DELETE FROM on all_review_data based on a data set containing the
records to delete.
}
\keyword{internal}
