% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_collect_data.R
\name{db_get_table}
\alias{db_get_table}
\title{Collect table from database}
\usage{
db_get_table(db_path, db_table = "all_review_data")
}
\arguments{
\item{db_path}{Character string. Path to the database.}

\item{db_table}{Character string. Name of the table to collect.}
}
\description{
Collect table from database
}
\keyword{internal}
