% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_save}
\alias{db_save}
\title{Append database table}
\usage{
db_save(data, db_path, db_table = "query_data")
}
\arguments{
\item{data}{A data frame.}

\item{db_path}{Character string with the file path to the database.}

\item{db_table}{Character vector with the name of the destination table that
needs to be appended.}
}
\value{
A table in a database will be appended. No values will be returned.
}
\description{
Saves a query to a table in the user database.
}
\examples{
db_save(mtcars, ":memory:", "mtcars_db")

}
\keyword{internal}
