% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_temp_connect}
\alias{db_temp_connect}
\title{Connect temporarily to a database.}
\usage{
db_temp_connect(db_path, code, drv = RSQLite::SQLite())
}
\arguments{
\item{db_path}{Path to the database.}

\item{code}{Code to execute with the temporary connection.}

\item{drv}{The DB driver to use. Standard the SQLite driver.}
}
\value{
Nothing will be returned by default.
}
\description{
Connect temporarily to a database.
}
\examples{
\dontrun{
 library(DBI)
 db_temp_connect(tempfile(), DBI::dbWriteTable(con, "test_table", mtcars))
}
 
}
\keyword{internal}
