% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_upgrade}
\alias{db_upgrade}
\title{Upgrade app database}
\usage{
db_upgrade(db_path)
}
\arguments{
\item{db_path}{Character vector. Path to the database}
}
\value{
A character vector indicating the version the DB was upgraded to. Run
for side effects on the database.
}
\description{
Helper function to upgrade the database when migrating to a newer version of
ClinSight that impacts the database.
}
