% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_data_fct_helpers.R
\name{delete_review_data}
\alias{delete_review_data}
\title{Delete review data}
\usage{
delete_review_data(review_df, latest_review_data, key_cols = key_columns)
}
\arguments{
\item{review_df}{Data frame containing old review data that need to be
updated.}

\item{latest_review_data}{Data frame containing latest review data.}

\item{key_cols}{A character vector containing the common key variables.
Defaults to \code{ClinSight} \code{\link[=key_columns]{key_columns()}}.}
}
\value{
A data frame containing only the rows to remove from the review data.
}
\description{
Identifies the records no longer in the dataset to be removed from
all_review_data.
}
\keyword{internal}
