% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_helpers.R
\name{dt_config}
\alias{dt_config}
\title{Configure DT helper}
\usage{
dt_config(data, table_name = "form")
}
\arguments{
\item{data}{A data frame used for display in a DT table. Number of rows will
be assessed}

\item{table_name}{character string, usually the form name}
}
\value{
list with three named objects: \code{dom}, \code{exts}, and \code{opts}
}
\description{
Small wrapper that helps handle some messiness preparing the correct \code{DT}
dom, extensions, & options when needed. Specifically, when & how to add an
Excel download button.
}
\keyword{internal}
