% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_helpers.R
\name{dt_options}
\alias{dt_options}
\title{Decide if Excel Button is appropriate}
\usage{
dt_options(data)
}
\arguments{
\item{data}{A data frame to use within the application.}
}
\description{
Small wrapper that helps handle some messiness preparing the correct \code{DT}
options when needed.
}
\keyword{internal}
