% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_figures.R
\name{fig_barplot}
\alias{fig_barplot}
\title{Bar chart}
\usage{
fig_barplot(data, x, col_fill, x_lab = "Adverse events", y_lab = "Individuals")
}
\arguments{
\item{data}{Data set to be used.}

\item{x}{Variable to plot.}

\item{col_fill}{Column with logical values, that defines the fill color.}

\item{x_lab}{Character string with the text on the x-axis.}

\item{y_lab}{Character string with the text on the y-axis.}
}
\value{
A ggplot2 object.
}
\description{
Creates a small bar chart.
}
\examples{
\dontrun{
 df <- mtcars |>
 dplyr::mutate(
  fill_color = sample(c(TRUE, FALSE), 1),
  .by = cyl
  )
fig_barplot(df, cyl, fill_color)
}
}
\keyword{internal}
