% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_figures.R
\name{fig_boxplots}
\alias{fig_boxplots}
\title{Create custom \code{ggplot()} boxplots.}
\usage{
fig_boxplots(
  data,
  xval = "site_code",
  yval = "item_value",
  text = "text_label",
  color_fill = "significance",
  sumval = NULL,
  title = NULL,
  x_title = "Site",
  y_title = NULL
)
}
\arguments{
\item{data}{data frame.}

\item{xval}{Character vector with the x-axis variable.}

\item{yval}{Character vector with the y-axis variable.}

\item{text}{Character vector, text label variable. Not used by ggplot2, but
can be used to display a text label on hover after converting the ggplot to
a plotly object with \code{\link[plotly:ggplotly]{plotly::ggplotly()}}.}

\item{color_fill}{Character vector with the name of the column containing the
categorical color scale.}

\item{sumval}{summary value, used to display a horizontal line}

\item{title}{Character vector, main figure title.}

\item{x_title}{x-axis title.}

\item{y_title}{y axis title.}
}
\description{
Creates a faceted \code{ggplot()} figure with boxplots per study site. Ignores
time/visit dependencies (will group data of all visits together).
}
\examples{
fig_data <- data.frame(
 "site_code" = sample(c("Site 1", "Site 2"), 50, replace = TRUE),
 "item_value" = runif(50, 1, 100),
 "subject_id"  = sample(1:10, 50, replace = TRUE),
 "text_label" = "test label",
 "significance" = "test"
 )
 fig_boxplots(
   fig_data,
   sumval = 20,
   title = "main_title",
   y_title = "y-axis_title",
   x_title = "x-axis_title"
   )
}
\keyword{internal}
