% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_figures.R
\name{fig_timeline}
\alias{fig_timeline}
\title{Create timeline}
\usage{
fig_timeline(data, min_events = 10)
}
\arguments{
\item{data}{Data frame to use.}

\item{min_events}{Minimum number of events to include in the compact
timeline. See also \code{\link[=fig_timeline]{fig_timeline()}}.}
}
\value{
A ggplot2 object.
}
\description{
Function to create a simple timeline figure using \code{ggplot2}.
}
