% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_config_fct_helpers.R
\name{filter_data}
\alias{filter_data}
\title{Filter app data}
\usage{
filter_data(data, sites, subject_ids, appdata)
}
\arguments{
\item{data}{A \code{Reactivevalues} object. filtered data will be written into
this object.}

\item{sites}{Character vector with sites to filter on.}

\item{subject_ids}{Character vector with all subject IDs. Used to keep the
correct order of subject IDs.}

\item{appdata}{Application data in long format, stored in a list. List
contains data frames named per form.}
}
\value{
A \code{reactivevalues} object.
}
\description{
Filter app data
}
