% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_appdata.R
\name{get_appdata}
\alias{get_appdata}
\title{Get Appdata}
\usage{
get_appdata(data, meta = metadata)
}
\arguments{
\item{data}{A data frame, commonly raw data merged with metadata using the
function \code{\link[=merge_meta_with_data]{merge_meta_with_data()}}.}

\item{meta}{A data frame with metadata. Used to retrieve all expected item
names and to retrieve the form types (continuous data, other data, ...).}
}
\value{
A list of data frames.
}
\description{
Converts data to a list of data frames split by \code{item_group}. It is currently
not needed to provide the application data in this format. However, the
function can be useful for clinsight data customization, especially in
combination with \code{\link[=create_table]{create_table()}}. It creates table classes of the
\code{item_group} name (in lower case letters, and all punctuation replaced by an
underscore), as long as there is a method available with the corresponding
item_group name for the function \code{\link[=create_table]{create_table()}}. If all data is of type
'continuous', then laboratory values such as lab limits and units are
cleaned.
}
