% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{get_db_connection}
\alias{get_db_connection}
\title{Connect to database}
\usage{
get_db_connection(
  path = db_path,
  drv = RSQLite::SQLite(),
  envir = parent.frame()
)
}
\arguments{
\item{path}{Character vector. Path to the database.}

\item{drv}{Database driver to use. Defaults to \code{RSQLite::SQLite()}.}

\item{envir}{Environment to make the connection available in.}
}
\value{
A database connection.
}
\description{
Small helper function to connect to database. This way, the connection
arguments are in one place, which makes it easier to change the database
driver in the entire app if needed.
}
\examples{
get_db_connection(tempfile(fileext = ".sqlite"))
}
