% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_helpers.R
\name{get_form_level_data}
\alias{get_form_level_data}
\title{Get form-level data.}
\usage{
get_form_level_data(data, all_forms, form_column = "item_group")
}
\arguments{
\item{data}{A data frame with form-level data. Should at least contain the
\code{form_column}.}

\item{all_forms}{A character vector containing the names of all forms for
which form-level data should be specified.}

\item{form_column}{Character string with the column in which the form names
are stored in \code{data}.}
}
\value{
A cleaned data frame with form-level data.
}
\description{
Internal function to clean form-level data and return a data frame with all
forms that should be specified, and include form-level data for all of them.
Will also set default values (as defined in the package) if the value is not
set and/or is missing.
}
