% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_figure_helpers.R
\name{get_ggplot_layer_names}
\alias{get_ggplot_layer_names}
\title{Get ggplot layer names}
\usage{
get_ggplot_layer_names(fig)
}
\arguments{
\item{fig}{A ggplot2 object.}
}
\value{
A character vector with all the layer names of the ggplot object.
}
\description{
Helper function to extract ggplot layer names from a ggplot object
}
\examples{
library(ggplot2)
fig <- ggplot(mtcars, aes(x = cyl, y = mpg, group = cyl)) + geom_boxplot() + geom_point()
get_ggplot_layer_names(fig)

}
\keyword{internal}
