% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_helpers.R
\name{get_meta_vars}
\alias{get_meta_vars}
\title{Get meta variables}
\usage{
get_meta_vars(data = appdata, meta = metadata)
}
\arguments{
\item{data}{List. Data to use.}

\item{meta}{List. metadata to use.}
}
\value{
a list with all important names to be used in a clinical trial.
}
\description{
Prepares and extracts consistent variable names based on both metadata
and the provided data. To be used in both the UI and server functions
of a Shiny app.
}
\keyword{internal}
