% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_helpers.R
\name{get_metadata}
\alias{get_metadata}
\title{Load metadata}
\usage{
get_metadata(
  filepath,
  expand_tab_items = c("common_forms", "study_forms", "general"),
  expand_cols = "suffix"
)
}
\arguments{
\item{filepath}{Filepath to the .xlsx file.}

\item{expand_tab_items}{Character vector with the names of the tabs of which
the items need to be expanded. If not empty, a new data frame will be
created named 'expanded_items', containing all items in the tabs of
\code{expand_tab_items}.}

\item{expand_cols}{Column names containing the columns for expansion. Will be
ignored if the variable \code{expand_tab_items} is left empty.}
}
\value{
A list with data frames.
}
\description{
Loads all sheets of an Excel file in a named list. Used for reading in
metadata.
}
