% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_appdata.R
\name{get_raw_data}
\alias{get_raw_data}
\title{Get raw data}
\usage{
get_raw_data(
  data_path,
  column_specs = metadata$column_specs,
  delim = ",",
  skip = 1
)
}
\arguments{
\item{data_path}{Path to the folder that contains the CSV files with the
application data.}

\item{column_specs}{A data frame containing at least the columns 'name_raw',
'name_new', and 'col_type'. Used to set the column specifications, and to
set the column names to the names expected for the application to function.}

\item{delim}{Delimiter to use to read in files.}

\item{skip}{Number of rows to skip when reading in files.}
}
\value{
A data frame with raw application data.
}
\description{
Get raw data
}
