% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_data_fct_helpers.R
\name{get_review_data}
\alias{get_review_data}
\title{Get review data}
\usage{
get_review_data(
  data,
  key_cols = key_columns,
  date_var = "event_date",
  edit_time_var = "edit_date_time"
)
}
\arguments{
\item{data}{A data frame, commonly raw data merged with meta data.}

\item{key_cols}{A character vector with the common key variable column names.
Assumes that these variables can identify a unique data point that needs to
be reviewed. Defaults to \code{ClinSight} \code{\link[=key_columns]{key_columns()}}.}

\item{date_var}{A character vector with the name of the date column.}

\item{edit_time_var}{A character vector with the name of the edit date-time
column.}
}
\value{
A data frame with review data.
}
\description{
Create a data frame containing all data to be reviewed. Selects the \emph{latest}
edit date-time for every event/item.
}
