% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_app_role_helpers.R
\name{get_valid_roles}
\alias{get_valid_roles}
\title{Get valid roles}
\usage{
get_valid_roles(roles)
}
\arguments{
\item{roles}{A character vector. Will be converted to lower case.}
}
\value{
A named character vector with the roles applicable for the
application and available for the current user.
}
\description{
Helper function to retrieve all roles that are relevant for the application.
}
