% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_authentication.R
\name{initialize_credentials}
\alias{initialize_credentials}
\title{Initialize database}
\usage{
initialize_credentials(
  credentials_db = "credentials_db.sqlite",
  credentials_pwd = Sys.getenv("DB_SECRET")
)
}
\arguments{
\item{credentials_db}{Character string. Path to the database.}

\item{credentials_pwd}{Password for the credentials database.}
}
\value{
Creates credentials database if needed. No values will be returned.
}
\description{
Creates a database with new credentials if it does not yet exist.
}
