% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{is_date}
\alias{is_date}
\title{Verify date class}
\usage{
is_date(x)
}
\arguments{
\item{x}{An atomic object (a vector).}
}
\value{
a logical of length one.
}
\description{
Small helper function to verify whether an object is of class Date or
POSIXct.
}
\examples{
\dontrun{
is_date(as.Date("2023-09-03"))
}
}
\keyword{internal}
