% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{key_columns}
\alias{key_columns}
\title{ClinSight key columns}
\format{
A character vector with the following key column names:
subject_id, event_name, item_group, form_repeat, item_name.
}
\usage{
key_columns
}
\description{
A character string containing the names of the columns that are needed to
distinguish unique records. All \code{ClinSight}-compatible data sets need to at
least contain these columns.
}
\keyword{datasets}
