% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_merge_categories.R
\name{merge_item_pairs_by_suffix}
\alias{merge_item_pairs_by_suffix}
\title{Merge items by suffix}
\usage{
merge_item_pairs_by_suffix(data, suffix_to_merge = "_ITEM_TO_MERGE_WITH_PAIR")
}
\arguments{
\item{data}{A data frame.}

\item{suffix_to_merge}{A character string with the suffix to merge.}
}
\value{
A data frame with the merged items.
}
\description{
Merges items with a suffix in their name with the item without the suffix.
}
