% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_appdata.R
\name{merge_meta_with_data}
\alias{merge_meta_with_data}
\title{Merge metadata with raw data}
\usage{
merge_meta_with_data(
  data,
  meta,
  expected_columns = c("lower_limit", "upper_limit", "unit", "significance",
    "reason_notdone")
)
}
\arguments{
\item{data}{A data set with raw study data.}

\item{meta}{A list of data frames containing metadata.}

\item{expected_columns}{A character vector with the columns that should be
expected in the data frame. If missing, these columns will be added with
missing data (thus, will be made explicitly missing).}
}
\value{
A data frame.
}
\description{
Study-specific function that will combine raw data gathered with
\code{\link[=get_raw_csv_data]{get_raw_csv_data()}} with study-specific metadata. It also fixes the metadata
suffix if needed, and renames the limits and significance values to the app
standard names. Some study-specific variables need to be created with this
step.
}
