% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Application Metadata}
\format{
\subsection{\code{metadata}: A list with 9 objects.}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{str(metadata)
#> List of 9
#>  $ column_names   : tibble [15 x 3] (S3: tbl_df/tbl/data.frame)
#>   ..$ name_raw: chr [1:15] "SiteCode" "SubjectId" "subjid" "data_page_id" ...
#>   ..$ name_new: chr [1:15] "site_code" "subject_id" "subj_id" "data_page_id" ...
#>   ..$ col_type: chr [1:15] "c" "c" "i" "i" ...
#>  $ events         : tibble [24 x 10] (S3: tbl_df/tbl/data.frame)
#>   ..$ event_id               : chr [1:24] "SCR" "C1D1" "C1D8" "C1D15" ...
#>   ..$ event_id_pattern       : chr [1:24] "^SCR$" "^C1D1$" "^C1D8$" "^C1D15$" ...
#>   ..$ is_regular_visit       : logi [1:24] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>   ..$ event_label_custom     : chr [1:24] NA NA NA NA ...
#>   ..$ event_name_custom      : chr [1:24] "Screening" "Cycle 1 Day 1" "Cycle 1 Day 8" "Cycle 1 Day 15" ...
#>   ..$ is_baseline_event      : logi [1:24] TRUE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ generate_labels        : logi [1:24] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ meta_event_order       : int [1:24] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ add_visit_number       : logi [1:24] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ add_event_repeat_number: logi [1:24] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ common_forms   :'data.frame':	307 obs. of  6 variables:
#>   ..$ item_name        : chr [1:307] "Day in Study" "Event Date" "Visit" "Lesion ID" ...
#>   ..$ item_type        : chr [1:307] "other" "other" "other" "other" ...
#>   ..$ item_form        : chr [1:307] "Efficacy" "Efficacy" "Efficacy" "Efficacy" ...
#>   ..$ form_level_id_var: chr [1:307] NA NA NA NA ...
#>   ..$ item_group       : chr [1:307] "Efficacy" "Efficacy" "Efficacy" "Efficacy" ...
#>   ..$ merge_with       : chr [1:307] NA NA NA NA ...
#>  $ study_forms    :'data.frame':	55 obs. of  9 variables:
#>   ..$ item_name        : chr [1:55] "Temperature" "RespiratoryRate" "PulseRate" "SystolicBloodPressure" ...
#>   ..$ item_type        : chr [1:55] "continuous" "continuous" "continuous" "continuous" ...
#>   ..$ item_form        : chr [1:55] "Vital Signs" "Vital Signs" "Vital Signs" "Vital Signs" ...
#>   ..$ form_level_id_var: chr [1:55] NA NA NA NA ...
#>   ..$ item_group       : chr [1:55] "Vital Signs" "Vital Signs" "Vital Signs" "Vital Signs" ...
#>   ..$ merge_with       : chr [1:55] NA NA NA NA ...
#>   ..$ unit             : chr [1:55] NA NA NA NA ...
#>   ..$ lower_limit      : chr [1:55] NA NA NA NA ...
#>   ..$ upper_limit      : chr [1:55] NA NA NA NA ...
#>  $ general        :'data.frame':	129 obs. of  6 variables:
#>   ..$ item_name        : chr [1:129] "Sex" "Age" "Childbearing Potential" "Reason" ...
#>   ..$ item_type        : chr [1:129] "other" "other" "other" "other" ...
#>   ..$ item_form        : chr [1:129] "General" "General" "General" "General" ...
#>   ..$ form_level_id_var: chr [1:129] NA NA NA NA ...
#>   ..$ item_group       : chr [1:129] "General" "General" "General" "General" ...
#>   ..$ merge_with       : chr [1:129] NA NA NA NA ...
#>  $ form_level_data:'data.frame':	19 obs. of  5 variables:
#>   ..$ item_group         : chr [1:19] "Efficacy" "Administration" "Adverse events" "Disease History" ...
#>   ..$ item_type          : chr [1:19] "other" "other" "other" "other" ...
#>   ..$ item_scale         : logi [1:19] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ use_unscaled_limits: logi [1:19] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>   ..$ review_required    : logi [1:19] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  $ table_names    : tibble [15 x 2] (S3: tbl_df/tbl/data.frame)
#>   ..$ table_name: chr [1:15] "Subject" "Visit" "Status" "Form Repeat" ...
#>   ..$ raw_name  : chr [1:15] "subject_id" "event_name" "status" "form_repeat" ...
#>  $ settings       :List of 6
#>   ..$ pre_merge_fns  : chr "arcus_pre_merge"
#>   ..$ pre_pivot_fns  : chr "arcus_pre_pivot"
#>   ..$ post_pivot_fns : chr "arcus_post_pivot"
#>   ..$ post_merge_fns : chr "arcus_post_merge"
#>   ..$ study_name     : chr "PRISM-1"
#>   ..$ treatment_label: chr "💊 Tₓ"
#>  $ items_expanded : tibble [958 x 11] (S3: tbl_df/tbl/data.frame)
#>   ..$ form_type        : chr [1:958] "common_forms" "common_forms" "common_forms" "common_forms" ...
#>   ..$ var              : chr [1:958] "Day in Study" "Event Date" "Timepoint" "Lesion Number" ...
#>   ..$ suffix           : chr [1:958] NA NA NA NA ...
#>   ..$ item_name        : chr [1:958] "Day in Study" "Event Date" "Visit" "Lesion ID" ...
#>   ..$ item_type        : chr [1:958] "other" "other" "other" "other" ...
#>   ..$ item_form        : chr [1:958] "Efficacy" "Efficacy" "Efficacy" "Efficacy" ...
#>   ..$ form_level_id_var: chr [1:958] NA NA NA NA ...
#>   ..$ item_group       : chr [1:958] "Efficacy" "Efficacy" "Efficacy" "Efficacy" ...
#>   ..$ unit             : chr [1:958] NA NA NA NA ...
#>   ..$ lower_limit      : chr [1:958] NA NA NA NA ...
#>   ..$ upper_limit      : chr [1:958] NA NA NA NA ...
}\if{html}{\out{</div>}}
}
}
\source{
Can be created with an Excel file. The Excel file format is chosen so
that the metadata can be changed easily per study. See
\code{raw-data/metadata.R} for details.
}
\usage{
metadata
}
\description{
A list of data frames and settings containing metadata that will be used for
the application. The metadata will be merged with with raw data. It controls
the variables that will be used in the application, and in which tab the
variables will be shown. The goal is that most, if not all, study-specific
data will be captured in the metadata, leaving the scripts to run the
application largely unaltered between studies.See \code{vignette("Metadata")} for
in-depth information.
}
\keyword{datasets}
