% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_db_synch_info.R
\name{mod_db_synch_info_ui}
\alias{mod_db_synch_info_ui}
\title{Database Synchronization information - Shiny module UI}
\usage{
mod_db_synch_info_ui(id)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module
Server.}
}
\description{
A shiny Module.
}
\details{
The module is currently initiated in the main sidebar. It shows the time at
which the database was synchronized last, and the time of the latest EDC data
entry (or data edit). In addition, a warning message will pop up if the
synchronization date is either unknown or older than one day.
}
\seealso{
\code{\link[=mod_db_synch_info_server]{mod_db_synch_info_server()}}
}
