% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_go_to_form.R
\name{mod_go_to_form_server}
\alias{mod_go_to_form_server}
\title{Go to form - Shiny module Server}
\usage{
mod_go_to_form_server(
  id,
  r,
  navinfo,
  navtable,
  tablerow,
  all_forms,
  form_name = "item_group",
  subject_id = "subject_id"
)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module Server.}

\item{r}{Common reactive values. Used to send back information about the
selected subject id (\code{r$subject_id}) to the main server. In addition,
\code{r$filtered_subjects} is needed to verify whether an ID selected in the \code{tablerow}
is in the active, filtered selection.}

\item{navinfo}{Common reactive values. Used to send back the request for the
change in form/subject id back to the server. Contains the character strings
\code{active_form}, \code{active_tab}, and \code{trigger_page_change}. The module will create
these variables if they are not yet available in \code{navinfo}.}

\item{navtable}{A reactive value. Table to look up the form and subject id to navigate to.}

\item{tablerow}{A reactive value. Table row associated with navtable to look
up the form and subject id in. For example, if \code{navtable} is converted
to a \code{\link[DT:datatable]{DT::datatable()}} object in \code{output$DTnavtable}, \code{tablerow} should be
\code{input$DTnavtable_table_rows_selected}. See also the
\href{https://rstudio.github.io/DT/shiny.html}{\code{DT::datatable()} documentation}.}

\item{all_forms}{A data frame containing two columns: a column named "form" with
all the form to navigate through in the correct order, and a column named
"main_tab", showing the name of the main_tab in which the form can be found.}

\item{form_name}{Character vector. Name of the column containing forms. If
defined, this column needs to be in the \code{navtable}. If this
variable is \code{NULL}, it will default to the first variable in vars$all_forms.}

\item{subject_id}{Character vector. Name of the column containing the subject
identifiers. This column needs to be defined in the \code{navtable}.}
}
\description{
A \code{shiny} module. Navigates to the subject and form that are mentioned
in a selected table row. Designed to be used with \code{\link[DT:datatable]{DT::datatable()}}, since
these tables can provide a reactive value with the \code{tablerow} highlighted on click.
}
\seealso{
\code{\link[=mod_go_to_form_server]{mod_go_to_form_server()}}
}
