% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_go_to_form.R
\name{mod_go_to_form_ui}
\alias{mod_go_to_form_ui}
\title{Go to form - Shiny module UI}
\usage{
mod_go_to_form_ui(id, button_text = "Go to form", class = "btn-warning m2")
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module Server.}

\item{button_text}{Character vector containing the text that will be displayed on the button.}

\item{class}{Class of the \code{actionButton}. Default class is "btn-warning m2".}
}
\description{
A \code{shiny} module. Navigates to the subject and form that are mentioned
in a selected table row. Designed to be used with \code{\link[DT:datatable]{DT::datatable()}}, since
these tables can provide a reactive value with the \code{tablerow} highlighted on click.
}
\seealso{
\code{\link[=mod_go_to_form_server]{mod_go_to_form_server()}}
}
