% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_header_widgets.R
\name{mod_header_widgets_ui}
\alias{mod_header_widgets_ui}
\title{Header widgets - Shiny module UI}
\usage{
mod_header_widgets_ui(id)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module
Server.}
}
\description{
A shiny module. Used to show user information of the active user in value
boxes on the top of the screen.
}
\details{
The value boxes in this module contain information about the active user.
They show the patient id, the patient status (active/inactive participant),
the number of adverse events (with a color code showing whether any new or
updated data is available), the number of forms that need a review, and a
timeline figure showing the number of visits that the patient performed. The
value box with adverse events also serves as a link to the adverse events
form. Furthermore, clicking on the box with forms to review will trigger
\code{\link[=mod_navigate_review_server]{mod_navigate_review_server()}}, opening a modal that shows the forms that need
review and the queries that are open of the active participant, to which you
can directly navigate to.
}
\seealso{
\code{\link[=mod_header_widgets_server]{mod_header_widgets_server()}}
}
