% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_sidebar.R
\name{mod_main_sidebar_ui}
\alias{mod_main_sidebar_ui}
\title{App sidebar - Shiny module UI}
\usage{
mod_main_sidebar_ui(id)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module
Server.}
}
\description{
A shiny module. Used to create the sidebar in the main application.
}
\details{
This sidebar contains the main controls for the application. In the sidebar,
a panel is available which allows the user to apply a review to the active
form and patient (forms can be marked as reviewed, a comment can be added,
and the review can be saved. In addition, it contains buttons to navigate to
the next/previous forms and to add a query to the active form.
}
\seealso{
\code{\link[=mod_main_sidebar_ui]{mod_main_sidebar_ui()}}
}
