% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_navigate_forms.R
\name{mod_navigate_forms_ui}
\alias{mod_navigate_forms_ui}
\title{Navigate forms - Shiny module UI}
\usage{
mod_navigate_forms_ui(id)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module Server.}
}
\description{
Shiny UI module containing the UI needed to navigate between
forms. The UI contains a \code{back} and \code{forward} which changes the active form.
In addition, the active form is displayed either in bold or in normal font
face, depending on whether the form contains any non-reviewed data.
}
\seealso{
\code{\link[=mod_navigate_forms_server]{mod_navigate_forms_server()}}
}
