% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_navigate_participants.R
\name{mod_navigate_participants_server}
\alias{mod_navigate_participants_server}
\title{Navigate participants - Shiny module Server}
\usage{
mod_navigate_participants_server(id, r, static_overview_data = NULL)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module Server.}

\item{r}{Common \code{reactiveValues}. Used to access \code{filtered_tables$General},
containing a data frame with general data to be displayed in the participant
selection modal.
In addition, it will be used to access the list of \code{filtered_subjects}
(character vector), and the currently active \code{subject_id} (character string).
The only parameter that the module will change, if requested by the user,
is \code{subject_id}.}
}
\description{
A \code{shiny} module. Used to show participant information in a
\code{\link[bslib:value_box]{bslib::value_box()}}. By clicking on the \code{\link[bslib:value_box]{bslib::value_box()}}, additional
participant information will be shown, as well as a selection menu to select
a different subject. Once the subject is changed, the active \code{subject_id} will
be changed in the application.
}
\seealso{
\code{\link[=mod_navigate_participants_ui]{mod_navigate_participants_ui()}} for the UI function
}
