% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_navigate_review.R
\name{mod_navigate_review_server}
\alias{mod_navigate_review_server}
\title{Navigate review tables - Shiny module Server}
\usage{
mod_navigate_review_server(
  id,
  r,
  rev_data,
  navinfo,
  all_forms,
  table_names = NULL
)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module Server.}

\item{r}{Common reactiveValues. Used to access \code{subjectid} (the active Subject),
to select the correct review data and to change the active subject id if
requested by the user.}

\item{rev_data}{Common reactiveValues. Used to access \code{rev_data$summary} (a data
frame with summarized review data), \code{rev_data$show_all} (logical, whether
to show data of everyone in the tables or only data of the selected subject id.)}

\item{navinfo}{Common reactive values. Used to send back the request for the
change in form/subject id back to the server. Contains the character strings
\code{active_form}, \code{active_tab}, and \code{trigger_page_change}. The module will create
these variables if they are not yet available in \code{navinfo}.}

\item{all_forms}{A data frame containing two columns: a column named "form" with
all the form to navigate through in the correct order, and a column named
"main_tab", showing the name of the main_tab in which the form can be found.
Used for navigation purposes, will be passed onto the internal Shiny module
\code{\link[=mod_go_to_form_server]{mod_go_to_form_server()}}.}

\item{table_names}{An optional character vector. If provided, will be used
within \code{\link[=datatable_custom]{datatable_custom()}}, to improve the column names in the final
interactive tables.}
}
\description{
A Shiny module. Used to display all data that needs review and all queries that are still open.
It enables the user to directly navigate to a specific form or patient.
}
\details{
The module will be called from the main server and will open a modal with
two tables: a table with forms that need to be reviewed and a table with
queries that are raised. There is a switch in the modal, which can toggle
the tables to only show data of the currently active and selected participant,
or to show data of all the participants that need to be reviewed.
There is a button to navigate to the form of a selected data frame. Once
clicked, this will activate the module \code{\link[=mod_go_to_form_server]{mod_go_to_form_server()}}, which is a
low-level helper module to change the active participant and active review
form to the one in the selected table row.
}
\seealso{
\code{\link[=mod_navigate_review_ui]{mod_navigate_review_ui()}}, \code{\link[=mod_go_to_form_ui]{mod_go_to_form_ui()}},
\code{\link[=mod_go_to_form_server]{mod_go_to_form_server()}}
}
