% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_navigate_review.R
\name{mod_navigate_review_ui}
\alias{mod_navigate_review_ui}
\title{Navigate review tables - Shiny module UI}
\usage{
mod_navigate_review_ui(id)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module Server.}
}
\description{
A shiny Module.
}
\details{
The module will be called from the main server and will open a modal with
two tables: a table with forms that need to be reviewed and a table with
queries that are raised. There is a switch in the modal, which can toggle
the tables to only show data of the currently active and selected participant,
or to show data of all the participants that need to be reviewed.
There is a button to navigate to the form of a selected data frame. Once
clicked, this will activate the module \code{\link[=mod_go_to_form_server]{mod_go_to_form_server()}}, which is a
low-level helper module to change the active participant and active review
form to the one in the selected table row.
}
\seealso{
\code{\link[=mod_navigate_review_server]{mod_navigate_review_server()}}
}
