% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_queries.R
\name{mod_queries_ui}
\alias{mod_queries_ui}
\title{Queries - Shiny module UI}
\usage{
mod_queries_ui(id)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module
Server.}
}
\description{
A Shiny module. Used to display all raised queries.
}
\details{
The module displays all raised queries in a table. Upon clicking on a query,
the query follow-up messages will be shown. In addition, it will be possible
to write a follow-up message. The logic for writing the follow-up message is
extracted in a different module, named \code{\link[=mod_query_follow_up_server]{mod_query_follow_up_server()}}. New
queries are created in the module \code{\link[=mod_query_add_server]{mod_query_add_server()}}. The button to
initiate this module can be found in the main sidebar
\code{\link[=mod_main_sidebar_server]{mod_main_sidebar_server()}}.
}
\seealso{
\code{\link[=mod_queries_server]{mod_queries_server()}}, \code{\link[=mod_go_to_form_ui]{mod_go_to_form_ui()}}, \code{\link[=mod_go_to_form_server]{mod_go_to_form_server()}}
}
