% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_query_add.R
\name{mod_query_add_server}
\alias{mod_query_add_server}
\title{Write Query - Shiny module Server}
\usage{
mod_query_add_server(id, r, active_form, db_path, available_data)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module
Server.}

\item{r}{Common \code{reactivevalues}. Used to access the active \code{subject_id} and
the current reviewer's name \code{user_name}.}

\item{active_form}{A reactive value, containing the active form.}

\item{db_path}{Character string with the path to the database.}

\item{available_data}{A data frame containing all available data, usually
created with the function \code{\link[=get_available_data]{get_available_data()}}.}
}
\description{
Shiny module. Used to create a new query and add it to the database. This
module can be activated by pressing an action button, which is currently
located in the applications' main sidebar (\code{\link[=mod_main_sidebar_server]{mod_main_sidebar_server()}}).
}
\details{
Upon activation of this module, a window will be shown with a fixed sidebar
and a field with the main content. The main content contains a field in which
a query can be written, contains  a \code{Save} and \code{Cancel} button, and shows the
name of the reviewer who is currently logged in. In the sidebar, information
is given about the currently active patient and form. In addition, drop-down
menus are shown to narrow down to which data the query is concerning. In
these menus, the applicable visit, and (optional) the applicable item within
the form can be selected.
}
\seealso{
\code{\link[=mod_query_add_ui]{mod_query_add_ui()}}, \code{\link[=mod_queries_ui]{mod_queries_ui()}}, \code{\link[=mod_queries_server]{mod_queries_server()}}
}
