% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_query_add.R
\name{mod_query_add_ui}
\alias{mod_query_add_ui}
\title{Write Query - Shiny module UI}
\usage{
mod_query_add_ui(id)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module
Server.}
}
\description{
Shiny module. Used to create a new query and add it to the database. This
module can be activated by pressing an action button, which is currently
located in the applications' main sidebar (\code{\link[=mod_main_sidebar_server]{mod_main_sidebar_server()}}).
}
\details{
Upon activation of this module, a window will be shown with a fixed sidebar
and a field with the main content. The main content contains a field in which
a query can be written, contains  a \code{Save} and \code{Cancel} button, and shows the
name of the reviewer who is currently logged in. In the sidebar, information
is given about the currently active patient and form. In addition, drop-down
menus are shown to narrow down to which data the query is concerning. In
these menus, the applicable visit, and (optional) the applicable item within
the form can be selected.
}
\seealso{
\code{\link[=mod_query_add_server]{mod_query_add_server()}}, \code{\link[=mod_queries_ui]{mod_queries_ui()}}, \code{\link[=mod_queries_server]{mod_queries_server()}}
}
