% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_query_follow_up.R
\name{mod_query_follow_up_server}
\alias{mod_query_follow_up_server}
\title{Add follow-up to a query - Shiny module Server}
\usage{
mod_query_follow_up_server(id, r, selected_query, db_path)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module Server.}

\item{r}{Common reactiveValues. Used to access the active \code{subject_id} and
the current reviewer's name \code{user_name}.}

\item{selected_query}{A reactive value. Contains a character vector with the
query ID, which is the unique identifier of each query.}

\item{db_path}{A character string with the path to the database.}
}
\description{
Add follow-up to a query - Shiny module Server
}
\seealso{
\code{\link[=mod_query_follow_up_server]{mod_query_follow_up_server()}}
}
