% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_query_follow_up.R
\name{mod_query_follow_up_ui}
\alias{mod_query_follow_up_ui}
\title{Add follow-up to a query - Shiny module UI}
\usage{
mod_query_follow_up_ui(id)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module Server.}
}
\description{
A shiny Module.
}
\seealso{
\code{\link[=mod_query_follow_up_server]{mod_query_follow_up_server()}}
}
