% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_report_server}
\alias{mod_report_server}
\title{Create Report - Shiny module Server}
\usage{
mod_report_server(id, r, rev_data, db_path, table_names)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module Server.}

\item{r}{Common reactiveValues. Used to access ....}

\item{rev_data}{Common reactiveValues. Used to access \code{rev_data$summary} (a data
frame with summarized review data) and \code{rev_data$show_all} (logical, whether
to show data of everyone in the tables or only data of the selected subject id.),
and \code{rev_data$show_modal} (to communicate that the review data modal should be opened).}

\item{db_path}{Character string with the path to the app database.}

\item{table_names}{An optional character vector. If provided, will be used
within \code{\link[=datatable_custom]{datatable_custom()}}, to improve the column names in the final
interactive tables.}
}
\description{
Shiny module. Used to create PDF reports of all review activities.
}
