% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report.R
\name{mod_report_ui}
\alias{mod_report_ui}
\title{Create Report - Shiny module UI}
\usage{
mod_report_ui(id)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module Server.}
}
\description{
Shiny module. Used to create PDF reports of all review activities.
}
\seealso{
\code{\link[=mod_report_server]{mod_report_server()}}
}
