% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_config.R
\name{mod_review_config_server}
\alias{mod_review_config_server}
\title{Review configuration - Shiny module Server}
\usage{
mod_review_config_server(id, r, app_data, sites, subject_ids)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module Server.}

\item{r}{Common reactiveValues. Used to pass on filtered data and filtered subjects
(based on selected sites/regions) to the main server. Expects to contain
\code{r$filtered_data}, \code{r$filtered_subjects} and \code{r$subject_id} (the '
active/current subject id'). The latter is needed because the \code{r$subject_id}
needs to be set to the first ID in the filtered selection to prevent a
non-selected subject_id to be active.}

\item{app_data}{List of data frames with the app data.}

\item{sites}{A data frame with columns "site_code", with all unique site
identifiers, and "region", the region of the study site.}

\item{subject_ids}{Character vector containing all subject ids. Used for
filtering subject ids. Using this string so that the correct order will be
retained. Dependency could be removed in future version.}
}
\description{
A Shiny module. Used to set the desired review configuration for the active
user.
}
\details{
When the module is called from the main server, it will show options to
select regions and sites of interest to the user. The user sets these
filters at the beginning of their session, to select the study sites that need
review. After selecting the appropriate filter settings and confirming the
review by clicking on the confirm review button, the internal data sets will
be filtered on the selected filter settings and only the selected
data will be shown in the application.
}
\seealso{
\code{\link[=mod_review_config_ui]{mod_review_config_ui()}}
}
