% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_config.R
\name{mod_review_config_ui}
\alias{mod_review_config_ui}
\title{Review configuration - Shiny module UI}
\usage{
mod_review_config_ui(id)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module Server.}
}
\description{
A Shiny module. Used to set the desired review configuration for the active
user.
}
\details{
When the module is called from the main server, it will show options to
select regions and sites of interest to the user. The user sets these
filters at the beginning of their session, to select the study sites that need
review. After selecting the appropriate filter settings and confirming the
review by clicking on the confirm review button, the internal data sets will
be filtered on the selected filter settings and only the selected
data will be shown in the application.
}
\seealso{
\code{\link[=mod_review_config_server]{mod_review_config_server()}}
}
