% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_form_tbl.R
\name{mod_review_form_tbl_ui}
\alias{mod_review_form_tbl_ui}
\title{Review forms table - Shiny module UI}
\usage{
mod_review_form_tbl_ui(id)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module
Server.}
}
\description{
Shiny module. Used to handle the logic and presentation of form review data.
}
\details{
This module handles the logic associated with the form review tables. It
handles all the logic associated with the table check boxes and the observers
needed to keep the browser table in-synch with the server table.
}
\seealso{
\code{\link[=mod_review_form_tbl_server]{mod_review_form_tbl_server()}}
}
