% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_forms.R
\name{mod_review_forms_server}
\alias{mod_review_forms_server}
\title{Review form - Shiny module Server}
\usage{
mod_review_forms_server(
  id,
  r,
  active_form,
  active_tab,
  review_required_data,
  db_path
)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module
Server.}

\item{r}{Common reactive values. Used to access subject_id, user_name,
review_data, and status_data. The module can change data in the review_data
and to status_data data frames.}

\item{active_form}{Reactive value. Used to determine which form is active and
whether that form needs review.}

\item{active_tab}{Reactive value containing the active tab. Needed to hide
the review controls if non-relevant tabs are selected (tabs that do not
contain common forms or study forms).}

\item{review_required_data}{a data frame with information about whether
review is mandatory for each form. Should contain the columns \code{item_group}
and \code{review_required}.}

\item{db_path}{Character string with the file path to the database.}
}
\description{
A \code{shiny} module. Used to mark data as reviewed and add comments
to data of a specific form/page in the database and in the internal
application data.

This module is used to display whether the form that is actively being
displayed in the application has any values that are newly entered or
updated since the last review session. In addition, if a page contains new
data, the user can mark all data in the form as being reviewed and can
(optionally) add a comment to this review action. The data will be saved in
a database. All review activity is stored with an audit-trail, with
date/time stamps and with the reviewer's name.
}
\seealso{
\code{\link[=mod_review_forms_ui]{mod_review_forms_ui()}}
}
