% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_start_page.R
\name{mod_start_page_server}
\alias{mod_start_page_server}
\title{Start page - Shiny module Server}
\usage{
mod_start_page_server(id, r, rev_data, navinfo, all_forms, table_names)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module Server.}

\item{r}{Common reactive values. Used to send back information about the
selected subject id (\code{r$subject_id}) to the main server. In addition,
\code{r$filtered_subjects} is needed to verify whether an ID selected in the \code{tablerow}
is in the active, filtered selection.}

\item{rev_data}{Common \code{reactiveValues()}. Used for navigation purposes, to
set the subject for which we want to browse the review data from.}

\item{navinfo}{Common reactive values. Used to send back the request for the
change in form/subject id back to the server. Contains the character strings
\code{active_form}, \code{active_tab}, and \code{trigger_page_change}. The module will create
these variables if they are not yet available in \code{navinfo}.}

\item{all_forms}{A data frame containing two columns: a column named "form" with
all the form to navigate through in the correct order, and a column named
"main_tab", showing the name of the main_tab in which the form can be found.}

\item{table_names}{An optional character vector. If provided, will be used
within \code{\link[=datatable_custom]{datatable_custom()}}, to improve the column names in the final
interactive tables.}
}
\description{
A Shiny module. Start page of the main application.
}
\details{
The start page contains navigation buttons and a table that shows one row per
patient. The table shows general study information such as study status
(active or inactive participant), main diagnosis, age, sex, number of
conducted study visits, and the patient's data status (new data, updated
data, or old data). In addition, in the module there are two navigation
options. First, the user can navigate to the selected patient directly.
Secondly, the user can indirectly call the module
\code{\link[=mod_navigate_review_server]{mod_navigate_review_server()}} to open a modal in which all the forms are
shown that need review of the selected patient. navigate to patient that is
currently highlighted. highlighted to a participant  study and options to
navigate directly to the selected forms/patients n the overview tables.
}
\seealso{
\code{\link[=mod_start_page_ui]{mod_start_page_ui()}}, \code{\link[=mod_go_to_form_ui]{mod_go_to_form_ui()}},
\code{\link[=mod_go_to_form_server]{mod_go_to_form_server()}}, \code{\link[=mod_navigate_review_server]{mod_navigate_review_server()}},
\code{\link[=mod_navigate_review_ui]{mod_navigate_review_ui()}}
}
