% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_start_page.R
\name{mod_start_page_ui}
\alias{mod_start_page_ui}
\title{Start page - Shiny module UI}
\usage{
mod_start_page_ui(id)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module Server.}
}
\description{
A Shiny module. Start page of the main application.
}
\details{
The start page contains navigation buttons and a table that shows one row per
patient. The table shows general study information such as study status
(active or inactive participant), main diagnosis, age, sex, number of
conducted study visits, and the patient's data status (new data, updated
data, or old data). In addition, in the module there are two navigation
options. First, the user can navigate to the selected patient directly.
Secondly, the user can indirectly call the module
\code{\link[=mod_navigate_review_server]{mod_navigate_review_server()}} to open a modal in which all the forms are
shown that need review of the selected patient. navigate to patient that is
currently highlighted. highlighted to a participant  study and options to
navigate directly to the selected forms/patients n the overview tables.
}
\seealso{
\code{\link[=mod_start_page_server]{mod_start_page_server()}}
}
