% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_study_forms.R
\name{mod_study_forms_server}
\alias{mod_study_forms_server}
\title{Study forms - Shiny module Server}
\usage{
mod_study_forms_server(
  id,
  form,
  form_data,
  form_review_data,
  form_items,
  active_subject,
  id_item = c("subject_id", "event_name", "item_group", "form_repeat", "item_name"),
  table_names = NULL,
  item_info
)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module
Server.}

\item{form}{A character string with the name of the form to display.}

\item{form_data}{A reactive value containing the study data of the respective
form.}

\item{form_review_data}{A reactive value containing the review data of the
respective form.}

\item{form_items}{A named character vector with the names of the expected
variables defined in the applicable \code{form}. Used in the UI to create a
filter with drop-down menu, to select the desired variables in a figure.
Used in the module Server to make sure that all expected columns are always
created, even if some variables are implicitly missing (which might occur
if there are not yet any values available for a specific variable). Also,
implicitly missing variables might give errors if part of the script relies
on the variables' presence. See also the parameter \code{expected_columns} in
\code{\link[=create_table.default]{create_table.default()}}.}

\item{active_subject}{A reactive value containing the active subject ID.}

\item{id_item}{Character vector containing the column names of the columns
that can uniquely identify one item/row.}

\item{table_names}{An optional character vector. If provided, will be used
within \code{\link[=datatable_custom]{datatable_custom()}}, to improve the column names in the final
interactive tables.}

\item{item_info}{A data frame containing the names of the study forms (in the
column \code{item_group}), and the columns \code{item_scale} \code{use_unscaled_limits},
which are used to customize the way the figures are shown in the page.}
}
\description{
A shiny module. Used to display figures and tables of a study-specific form
and will be accessible in through the study form tab on the main page.
}
\details{
The module displays tables and figures, and shows data of the currently
active subject. In the figures, the data can easily be compared with data of
other study participants. Data that is new or updated will have large points,
so that the reviewer can focus on new data. By default, all items of a form
will be shown, but the reviewer can select single or multiple variables to
focus on in the figure. There is also an option to switch from graphical view
to table view. In a table view, the same data will be shown in wide-table
format, with new/updated data shown in bold. If the values contain units, the
original units will be shown. The \verb{Study forms} module is used in the main
\code{app_server} to create all applicable study form pages.
}
\seealso{
\code{\link[=mod_study_forms_ui]{mod_study_forms_ui()}}, \code{\link[=mod_review_form_tbl_ui]{mod_review_form_tbl_ui()}},
\code{\link[=mod_review_form_tbl_server]{mod_review_form_tbl_server()}}
}
