% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_study_forms.R
\name{mod_study_forms_ui}
\alias{mod_study_forms_ui}
\title{Study forms - Shiny module UI}
\usage{
mod_study_forms_ui(id, form, form_items)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module
Server.}

\item{form}{A character string with the name of the form to display.}

\item{form_items}{A named character vector with the names of the expected
variables defined in the applicable \code{form}. Used in the UI to create a
filter with drop-down menu, to select the desired variables in a figure.
Used in the module Server to make sure that all expected columns are always
created, even if some variables are implicitly missing (which might occur
if there are not yet any values available for a specific variable). Also,
implicitly missing variables might give errors if part of the script relies
on the variables' presence. See also the parameter \code{expected_columns} in
\code{\link[=create_table.default]{create_table.default()}}.}
}
\description{
Shiny module. Used to display common form data in the dedicated tab.
}
\details{
The module displays tables of common form data in a wide format. Included
common forms are currently: \verb{Adverse events}, \verb{Medical History},
\code{Medication}, and \verb{Conc. Procedures}. The applicable forms can be flexibly
changed in the metadata. The tables shown are overview tables in wide format,
similar to the ones in \code{\link[=mod_study_forms_server]{mod_study_forms_server()}}. When the common form
\verb{Adverse events} is selected, the module will show an additional table with
Severe Adverse Events above the table with Adverse Events.
The \verb{common forms} module is used in the main server to create all applicable
common form pages.
}
\seealso{
\code{\link[=mod_study_forms_server]{mod_study_forms_server()}}
}
