% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{order_string}
\alias{order_string}
\title{Order string}
\usage{
order_string(string = appdata$`CBC regular`$subject_id)
}
\arguments{
\item{string}{Character string to be used.}
}
\value{
An ordered string.
}
\description{
Small helper function to define the proper order per individual/site, to
arrange the sites/individuals properly within the application.
Arranges first on all string letters, and then on the numbers.
}
\examples{
order_string(rownames(mtcars))

}
