% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_figures.R
\name{plotly_figure}
\alias{plotly_figure}
\title{Plotly figure}
\usage{
plotly_figure(data, fig = "timeseries_fig", height = NULL, ...)
}
\arguments{
\item{data}{Data frame to use.}

\item{fig}{A character vector with the figure function to use.}

\item{height}{Height of the plot.}

\item{...}{Other variables, passed onto the figure functions that are called to create the plots.}
}
\description{
Makes a ggplot2 figure interactive using plotly
}
\seealso{
\code{\link[=fig_timeseries]{fig_timeseries()}}, \code{\link[=fig_boxplots]{fig_boxplots()}}
}
