% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_figure_helpers.R
\name{remove_boxplot_outliers}
\alias{remove_boxplot_outliers}
\title{Remove boxplot outliers}
\usage{
remove_boxplot_outliers(fig)
}
\arguments{
\item{fig}{A \code{plotly} object created with \code{\link[plotly:ggplotly]{plotly::ggplotly()}}.}
}
\value{
The same \code{\link[plotly:ggplotly]{plotly::ggplotly()}} object, now with outliers in boxplots deleted.
}
\description{
Function to remove boxplot outliers and hover info from a \code{\link[plotly:ggplotly]{plotly::ggplotly()}}
object.
Only boxplot layers will be altered. Works also in faceted \code{ggplotly} figures.
This is needed since boxplot outliers will cannot be removed in interactive
\code{plotly} figures due to a bug. See also
}
\examples{
\dontrun{
library(ggplot2)
library(plotly)
set.seed(2023) # for reproducible jitter
fig <- ggplotly({
  ggplot(iris, aes(factor(1), Petal.Width)) + geom_boxplot(outlier.shape = NA) + 
    geom_jitter(col = "red") + 
    facet_wrap(~Species)
 })
 # Outliers are visible in black in the plotly figure, even though they should not:
fig

#The function `remove_boxplot_outliers()` corrects for this:
remove_boxplot_outliers(ggplotly(fig))
}

}
\keyword{internal}
