% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  data_folder = NULL,
  credentials_pwd = Sys.getenv("DB_SECRET"),
  test_mode = FALSE,
  ...
)
}
\arguments{
\item{onStart}{A function that will be called before the app is actually run.
This is only needed for \code{shinyAppObj}, since in the \code{shinyAppDir}
case, a \code{global.R} file can be used for this purpose.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}

\item{data_folder}{Character string. The folder in which all data resides is
usually set in the config.yml file. However, this can be overwritten if a
path is set in this argument. Useful for testing purposes.}

\item{credentials_pwd}{Character string with the credentials' database
password.}

\item{test_mode}{Logical, whether to run the application in test mode.}

\item{...}{arguments to pass to golem_opts. See \code{?golem::get_golem_options}
for more details.}
}
\description{
Run the Shiny Application
}
\details{
There are several elements defined in \code{golem-config.yml} that require
configuration before launching the application for the first time. To name a
few:
\itemize{
\item \code{user_db} a Character string providing the path to the app databases.
If it does not exist, one will be created based on app data and metadata,
with all data labeled as new'/not yet reviewed.
\item \code{credentials_db} Character string. Path to the credentials database.
}

The other two elements are \code{meta_data} and \code{study_data}, accepting file paths
to the app's primary source of data, stored as RDS files. As such, below are
comprehensive data specifications for these objects.
\subsection{\code{study_data}}{

The RDS file ported to the \code{study_data} element is a data.frame containing the
following required columns:
\itemize{
\item \code{site_code}: character or integer, identifier for study site; If an integer,
recommended to add prefix "Site" as this will display more intuitively in
the application's UI
\item \code{subject_id}: character, unique identifier for a subject
\item \code{event_repeat}: integer, helps keep track of unique \code{event_id} for a single
\code{subject_id} and \code{event_date}
\item \code{event_id}: character, names that help classify types of  \code{event_name}s
into like-groups, generally characterized by site visits. For example,
"SCR" for the screening visit, "VIS" for Visit X (where X is some integer),
and "EXIT" for when the patient exits the study trial. However, some
\code{event_id}s track events that could apply outside of any visit, like AE,
ConMed, Medical History, etc.
\item \code{event_name}: character, an "event" generally characterizes some sort of
site visit, whether that be a "Screening", "Visit X" (where X is some
integer), "Exit", or "Any Visit".
\item \code{event_date}: Date, the date associated with \code{event_name}
\item \code{form_id}: character, a unique identifier for the form the \code{item_name} metric
and \code{item_value} were pulled from. Note: when \code{item_type} is continuous,
\code{form_id} can contain several different \code{item_group}s. However, when
\code{item_type} is 'other', \code{item_group} can be made up of several \code{form_id}
values.
\item \code{form_repeat}: integer, helps keep track of unique \code{item_name}s collected
from a specific \code{form_id} for a given \code{subject_id}. \code{form_repeat} is
particularly helpful when conslidating data like Adverse Events into this
data format. Specifically, if more than one AE is collected on a patient,
they'll have more than one \code{form_repeat}
\item \code{edit_date_time}: datetime (POSIXct), the last time this record was edited
\item \code{db_update_time}: datetime (POSIXct), the last time the database storing this
record was updated.
\item \code{region}: character, describing the region code that \code{site_code} falls under
\item \code{day}: a difftime number, meaning it contains both a number and unit of
time. It measures the number of days each visit is from screening
\item \code{vis_day}: numeric, a numeric representation of \code{day}
\item \code{vis_num}: numeric, a numeric representation of \code{event_name}
\item \code{event_label}: character, an abbreviation of \code{event_name}
\item \code{item_name}:  character, describes a metric or parameter of interest.
\item \code{item_type}:  character, classifies \code{item_name}s into either 'continuous'
or 'other', where continuous types are those generally associated with the
CDISC "basic data structure" (BDS). That is, each \code{item_name} metric is
collected over time at a patient visit (\code{event_name}). The 'other' type
represents all non-time dependent measures, like demographic info, adverse
events, Medications, medical history, etc.
\item \code{item_group}:  character, provides is a high level category that groups
like-\code{item_name}s together. For example, and \code{item_group} = 'Vital Signs'
will group together pertinent \code{item_name} metrics like BMI, Pulse, Blood
pressure, etc.
\item \code{item_value}:  character, the measurement collected for a given \code{item_name}.
The value collected may be a number like 150 (when collecting a patient's
weight) or a word (such as 'white' for the subject's race).
\item \code{item_unit}:  character, tracking the unit of measurement for \code{item_name}
and \code{item_value}.
\item \code{lower_lim}: numeric, some \code{item_name}s (particularly the 'continuous' type)
have a pre-defined range of values that are considered normal. This is the
lower limit to that range.
\item \code{upper_lim}: numeric, some \code{item_name}s (particularly the 'continuous' type)
have a pre-defined range of values that are considered normal. This is the
upper limit to that range.
\item \code{significance}:  character, either 'CS' which means 'Clinically Significant'
or 'NCS' which means 'Not Clinically Significant'
\item \code{reason_notdone}:  character, an effort to describe why the \code{item_value}
field is \code{NA} / missing.
}
}

\subsection{\code{meta_data}}{

The RDS file ported to the \code{meta_data} configuration is a list of data.frames
which (not surprisingly) contains metadata used for the application. Prior to
launching the application, the metadata will be merged with the \code{study_data}.
to dictate which variables will be included in the application, and in which
tab the variables will be displayed. The goal is that most, if not all
study-specific data will be captured in the metadata, leaving the scripts to
run the application largely unaltered between studies.

Specifications for the list of data.frames include:
:
\itemize{
\item \code{events}: Used to create a simple timeline in the application, with
predefined number of planned visits, N. It contains the following columns:
\itemize{
\item \code{event_number}: integer. Example: 0, 1, 2, ..., N
\item \code{event_name}: character. Example: "Screening", "Visit 1", "Visit 2",
..., "Visit N"
\item \code{event_label}: character. Example: "V0", "V1", "V2", ..., "VN"
}
\item \code{common_forms}: Used to select and rename the variables of interest in
the raw data when transformed into the desired \code{study_data} format. Note:
creating the \code{study_data} data.frame should use \code{merge_meta_with_data()}
where (not surprisingly), the metadata is merged with the raw study data.
\code{common_forms} contains the columns below:
\itemize{
\item \code{var}: character, the variable name to display in the table, mapped
from a known \code{item_name} provided in \code{study_data}. Example:
\code{item_name = "AE Name"} will be replaced by "AE_AETERM" when
\code{var = "AE_AETERM"}.
\item \code{suffix}: Usually blank in this data.frame. This column is more
commonly used in the \code{study_forms} data.frame
\item \code{item_name}: character, known \code{item_name}s found in \code{study_data}. There
are certain \code{item_name}s that are required, even if missing in
\code{study_data}, including: "AE Name", "AE start date", "AE end date", "AE
date of worsening", "AE CTCAE severity", "AE CTCAE severity worsening",
"Serious Adverse Event", and "SAE Start date".
\item \code{item_type}: character, known \code{item_type} corresponding to those found
for \code{item_name}s in \code{study_data}.
\item \code{item_group}: character, known \code{item_group} corresponding to those found
for \code{item_name}s in \code{study_data}.
}
\item \code{study_forms}: Contains the same columns as the data.frame \code{common_forms},
and in addition the columns \code{unit}, \code{lower_limit}, \code{upper_limit}. Used to
select and rename the raw data variables of interest. In addition, the
\code{suffix} column is used more regularly in this "study" context. This is because,
these variable names may have a consistent trunk / stem, with
varying suffixes to describe a similar style measurement. So instead of creating a
new row for these variables in the \code{meta_data} data.frame, we allow for inclusion of several
suffixes. For example, VS_PULSE measures beats/min. Typically, these measures
are collected using VS_PULSE_VSORRES & VS_PULSE_VSREAND, but in this format,
we can list the stem "VS_PULSE" as the \code{var} and \code{"VSORRES, VSREAND"} in the suffix field.
As for the new columns, they are defined as follows:
\itemize{
\item \code{unit}: character, unit of measure
\item \code{lower_limit}: numeric, the lower limit of what's considered clinically significant
\item \code{upper_limit}: numeric, the upper limit of what's considered clinically significant
}
\item \code{general}: Contains the same columns as \code{common_forms} and is used in the
same way. That is, it's used to select and rename the raw data when
transformed into the desired \code{study_data} format. Note:
creating the \code{study_data} data.frame should use \code{merge_meta_with_data()}
where (not surprisingly), the metadata is merged with the raw study data.
Please refer to the \code{common_forms} spec above. However,
I will note that there are certain \code{item_name}s that are required, even if
missing in \code{study_data}, including: "Age", "Sex", "ECOG", "Eligible",
"WHO.classification", "DiscontinuationReason", "DrugAdminDate", and
"DrugAdminDose".
\item \code{groups}: Contains the columns \code{item_group}, \code{item_type}, \code{item_scale},
\code{use_unscaled_limits}.
}
}
}
