% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_report_fct_helpers.R
\name{select_report_data}
\alias{select_report_data}
\title{Select report data}
\usage{
select_report_data(
  data,
  report_table,
  report_type,
  report_reviewer,
  include_from_date = input$include_from_date
)
}
\arguments{
\item{data}{A data frame with the report data.}

\item{report_table}{Character vector. Can be either "review_data" or "query_data".}

\item{report_type}{Character vector. Controls which data will be included in the report.
Either "session" (only data from the current session and review will be included),
or "all" (all review data will be included).}

\item{report_reviewer}{Character vector with the name of the current reviewer.
Used to filter the data. Will be ignored if report_type is not "session".}

\item{include_from_date}{Character vector with the date from which data
should be included. Will be ignored if report_type is not "session".}
}
\value{
A data frame with the required report data.
}
\description{
Helper function for mod_report. Filters the correct report data and renames
the columns to a more report-friendly name.
}
