% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_config_fct_helpers.R
\name{set_user_role}
\alias{set_user_role}
\title{Set user role}
\usage{
set_user_role(data, user_role)
}
\arguments{
\item{data}{A \code{Reactivevalues} object. filtered data will be written into
this object.}

\item{user_role}{Character vector with the selected user role.}
}
\description{
Sets user role in a \code{reactiveValues()} object. Designed to be used within
\code{\link[=mod_review_config_server]{mod_review_config_server()}}.
}
