% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{simplify_string}
\alias{simplify_string}
\title{Simplify String}
\usage{
simplify_string(x)
}
\arguments{
\item{x}{a character vector or a factor. Factors will be silently converted
to character strings.}
}
\value{
character vector of the same length as the input vector.
}
\description{
Converts readable names to clean ids that are easier to handle in R.
Replaces all punctuation and spaces for an underscore, and transforms all
uppercase letters to lowercase.
}
\examples{
simplify_string(c(" Dirty_name. to Clean.#$", "#$another   complicated. Name"))
}
