% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_data_fct_helpers.R
\name{summarize_review_data}
\alias{summarize_review_data}
\title{Summarize review data}
\usage{
summarize_review_data(
  data,
  key_cols = c("subject_id", "item_group"),
  event_var = "event_name",
  date_time_vars = "edit_date_time",
  status_var = "status",
  collapse_exclude = "Any visit"
)
}
\arguments{
\item{data}{Data frame with review data.}

\item{key_cols}{A character vector containing the common key variables.}

\item{event_var}{A character string containing the name of the column with
the event names.}

\item{date_time_vars}{A character vector with the names of the date-time
variables. Can be multiple variable names.}

\item{status_var}{A character vector with the name of the status column.}

\item{collapse_exclude}{A character vector with the values to be excluded
from the column collapse.}
}
\value{
A data frame. Required columns are: subject_id, item_group,
event_name, edit_date_time, status, reviewed.
}
\description{
Creates a summary data frame based on review data. Creates one row per
subject id and form. All events (or visits) of a specific subject id and form
will be collapsed into one column separated with a comma, using
\code{\link[=collapse_column_vals]{collapse_column_vals()}} internally. Used to summarize data in the server,
which is then passed to the modules \code{\link[=mod_header_widgets_server]{mod_header_widgets_server()}},
\code{\link[=mod_navigate_review_server]{mod_navigate_review_server()}}, and \code{\link[=mod_report_server]{mod_report_server()}}.
}
\keyword{internal}
