% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{time_stamp}
\alias{time_stamp}
\title{Time stamp}
\usage{
time_stamp(form = "\%Y-\%m-\%d \%H:\%M:\%S \%Z", timezone = "UTC")
}
\arguments{
\item{form}{Format to be used for the time stamp.
See also \code{\link[base:strptime]{base::strptime()}} for the possible formats.}

\item{timezone}{Timezone to use.}
}
\value{
A character string with time stamp in it.
}
\description{
Simple wrapper around Sys.time to get a time stamp in
unified timezone and format.
}
\examples{
\dontrun{
time_stamp()
}
}
\keyword{internal}
