% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{title_case}
\alias{title_case}
\title{Capitalize words}
\usage{
title_case(x)
}
\arguments{
\item{x}{Character string.}
}
\value{
A character string.
}
\description{
Convert sentence to upper case to capital words

Adapted from the base R toupper() function help page.
}
\examples{
title_case("this will be converted to upper case")
# [1] "This Will Be Converted To Upper Case"

}
