% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_form_helpers.R
\name{update_tbl_data_from_datatable}
\alias{update_tbl_data_from_datatable}
\title{Update Server Table from Selection}
\usage{
update_tbl_data_from_datatable(tbl_data, review_selection)
}
\arguments{
\item{tbl_data}{A data frame containing the server table.}

\item{review_selection}{The review selection data frame input from the
datatable.}
}
\value{
A data frame containing the updated table data.
}
\description{
Updates the server table object based on the user selection.
}
