% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.R
\name{vector_select}
\alias{vector_select}
\title{Select from vector}
\usage{
vector_select(x, include = NULL, exclude = NULL)
}
\arguments{
\item{x}{A character vector.}

\item{include}{Optional. If provided, will only include vector elements that
match this regular expression.}

\item{exclude}{Optional. If provided, will exclude vector elements that match
this regular expression.}
}
\value{
A character vector.
}
\description{
Small helper function to select values from a character vector, based on
regular expressions.
}
\examples{
vector_select(names(iris), include = "Sepal", exclude = "Width")

}
\keyword{internal}
