% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_helpers.R
\name{add_timevars_to_data}
\alias{add_timevars_to_data}
\title{Add time vars to raw data}
\usage{
add_timevars_to_data(data, events)
}
\arguments{
\item{data}{A data frame}

\item{events}{A data frame with events. Usually contains the columns
\code{event_id}, \code{event_id_pattern}, \code{is_regular_visit}, \code{event_label_custom},
\code{event_name_custom}, and \code{add_sequence_to_name}.}
}
\value{
A data frame, with derivative time and event variables, needed for
ClinSight to function properly.
}
\description{
Add time vars to raw data
}
\keyword{internal}
