% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_helpers.R
\name{add_timevars_to_data}
\alias{add_timevars_to_data}
\title{Add time vars to raw data}
\usage{
add_timevars_to_data(
  data,
  events,
  label_type = c("hybrid", "derived", "original")
)
}
\arguments{
\item{data}{A data frame}

\item{events}{A data frame with events. Needs at least the columns
\code{visit_name}, \code{visit_suffix}, \code{event_id_pattern}, \code{is_visit_day}, and
\code{max_n_events}.}

\item{label_type}{Character vector to control the type of label set. Variable
not yet in use.}
}
\value{
A data frame, with derivative time and event variables, needed for
ClinSight to function properly.
}
\description{
Add time vars to raw data
}
\keyword{internal}
