% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_customize_events.R
\name{add_timevars_to_data}
\alias{add_timevars_to_data}
\title{Add time vars to raw data}
\usage{
add_timevars_to_data(data, events)
}
\arguments{
\item{data}{A data frame}

\item{events}{A data frame with events. Expected to be cleaned with the
function \code{\link[=clean_event_metadata]{clean_event_metadata()}}.}
}
\value{
A data frame, with derivative time and event variables, needed for
ClinSight to function properly.
}
\description{
Add time vars to raw data
}
\seealso{
\code{\link[=clean_event_metadata]{clean_event_metadata()}}
}
\keyword{internal}
