% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_authentication.R
\name{authenticate_server}
\alias{authenticate_server}
\title{Authenticate server}
\usage{
authenticate_server(
  test_mode = FALSE,
  sites = app_vars$Sites$site_code,
  roles = c("Medical Monitor", "Data Manager", "Administrator", "Investigator"),
  credentials_db = app_sys("app/www/credentials_db.sqlite"),
  credentials_pwd = Sys.getenv("DB_SECRET")
)
}
\arguments{
\item{test_mode}{Logical, whether to start the application in test mode.}

\item{sites}{Character vector. Study sites that can be allocated to a user.}

\item{roles}{Character vector. Roles that can be allocated to a user.}

\item{credentials_db}{Character vector. Path to the credentials
database.}

\item{credentials_pwd}{Character vector, containing the database
password.}
}
\description{
Function to authenticate the main server.
}
