% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_authentication.R
\name{authenticate_ui}
\alias{authenticate_ui}
\title{Authenticate UI}
\usage{
authenticate_ui()
}
\description{
Authentication implementation in the UI, using \code{shinymanager}.
}
\keyword{internal}
