% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clinsightful_data}
\alias{clinsightful_data}
\title{Clinical Trial test data}
\format{
a data.frame with 6,483 rows and 24 variables.
}
\source{
Created with \code{data-raw/create_random_data.R}
}
\usage{
clinsightful_data
}
\description{
A data.frame containing randomly created clinical trial data. Used for
testing purposes. It will also be used to run the app with example data when
executing \code{run_app()} with the default configuration settings, or when
running \code{golem::run_dev()}.
}
\keyword{datasets}
