% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_db_collect_data.R
\name{collect_query_data}
\alias{collect_query_data}
\title{Collect query data}
\usage{
collect_query_data(
  db_path,
  db_table = "query_data",
  time_var = "timestamp",
  query_id = "query_id",
  query_follow_up_number = "n",
  resolved_var = "resolved",
  resolved_date_var = "resolved_date"
)
}
\arguments{
\item{db_path}{A character string with the file path to the database.}

\item{db_table}{Character string. Name of the table to collect. Will only be
used if \code{data} is a character string to a database.}

\item{time_var}{Character string with the name of the timestamp variable.}

\item{query_id}{Character string with the name of the query id variable}

\item{query_follow_up_number}{Character string. Follow-up number of the
query. Will be incremented for follow-up message on the initial query.}

\item{resolved_var}{Character string with the name of the variable that shows
whether the query is resolved ('Yes') or not ('No').}

\item{resolved_date_var}{Character string. Date-time when the query was
marked as resolved.}
}
\value{
A data frame collected from a remote database.
}
\description{
Collects query data from a remote database.
}
