% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs_app_data}
\alias{cs_app_data}
\title{Application Data}
\format{
\subsection{\code{cs_app_data}: A list with 11 objects.}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{str(cs_app_data)
#> List of 11
#>  $ Adverse events  : advrs_vn [719 x 22] (S3: adverse_events/common_forms/tbl_df/tbl/data.frame)
#>   ..$ site_code     : chr [1:719] "BEL04" "BEL04" "BEL04" "BEL04" ...
#>   ..$ subject_id    : chr [1:719] "BEL_04_772" "BEL_04_772" "BEL_04_772" "BEL_04_772" ...
#>   ..$ event_id      : chr [1:719] "COMMON_AE" "COMMON_AE" "COMMON_AE" "COMMON_AE" ...
#>   ..$ event_date    : Date[1:719], format: "2023-08-31" "2023-08-31" ...
#>   ..$ event_repeat  : int [1:719] 1 1 1 1 1 1 1 1 2 2 ...
#>   ..$ form_id       : chr [1:719] "AE" "AE" "AE" "AE" ...
#>   ..$ form_repeat   : int [1:719] 1 1 1 1 1 1 1 1 2 2 ...
#>   ..$ edit_date_time: POSIXct[1:719], format: "2023-08-31 12:09:00" "2023-08-31 12:09:00" ...
#>   ..$ day           : 'difftime' num [1:719] 87 87 87 87 ...
#>   .. ..- attr(*, "units")= chr "days"
#>   ..$ event_name    : chr [1:719] "Any visit" "Any visit" "Any visit" "Any visit" ...
#>   ..$ event_label   : Factor w/ 8 levels "SCR","V1","V2",..: NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ form_type     : chr [1:719] "common_forms" "common_forms" "common_forms" "common_forms" ...
#>   ..$ item_name     : chr [1:719] "AE Number" "AE Name" "AESI" "AE start date" ...
#>   ..$ item_type     : chr [1:719] "other" "other" "other" "other" ...
#>   ..$ item_group    : chr [1:719] "Adverse events" "Adverse events" "Adverse events" "Adverse events" ...
#>   ..$ item_unit     : chr [1:719] "(unit missing)" "(unit missing)" "(unit missing)" "(unit missing)" ...
#>   ..$ lower_lim     : num [1:719] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ upper_lim     : num [1:719] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ item_value    : chr [1:719] "1" "Hypotension" "None" "2023-07-07" ...
#>   ..$ significance  : chr [1:719] NA NA NA NA ...
#>   ..$ reason_notdone: chr [1:719] NA NA NA NA ...
#>   ..$ region        : chr [1:719] "BEL" "BEL" "BEL" "BEL" ...
#>   ..- attr(*, "synch_time")= chr "2023-09-15 10:10:00 UTC"
#>  $ CBC regular     : continus [381 x 25] (S3: continuous/tbl_df/tbl/data.frame)
#>   ..$ site_code     : chr [1:381] "BEL04" "BEL04" "BEL04" "BEL04" ...
#>   ..$ subject_id    : chr [1:381] "BEL_04_133" "BEL_04_133" "BEL_04_133" "BEL_04_133" ...
#>   ..$ event_id      : chr [1:381] "SCR" "SCR" "SCR" "SCR" ...
#>   ..$ event_date    : Date[1:381], format: "2023-06-05" "2023-06-05" ...
#>   ..$ event_repeat  : int [1:381] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ form_id       : chr [1:381] "LBHM" "LBHM" "LBHM" "LBHM" ...
#>   ..$ form_repeat   : int [1:381] 1 1 1 1 1 1 2 2 2 2 ...
#>   ..$ edit_date_time: POSIXct[1:381], format: "2023-07-05 12:29:00" "2023-07-05 12:29:00" ...
#>   ..$ day           : 'difftime' num [1:381] 0 0 0 0 ...
#>   .. ..- attr(*, "units")= chr "days"
#>   ..$ event_name    : chr [1:381] "Screening" "Screening" "Screening" "Screening" ...
#>   ..$ event_label   : Factor w/ 8 levels "SCR","V1","V2",..: 1 1 1 1 1 1 2 2 2 2 ...
#>   ..$ form_type     : chr [1:381] "study_forms" "study_forms" "study_forms" "study_forms" ...
#>   ..$ item_name     : Factor w/ 6 levels "Haemoglobin",..: 4 5 6 1 2 3 4 5 6 1 ...
#>   ..$ item_type     : chr [1:381] "continuous" "continuous" "continuous" "continuous" ...
#>   ..$ item_group    : chr [1:381] "CBC regular" "CBC regular" "CBC regular" "CBC regular" ...
#>   ..$ item_unit     : chr [1:381] "10^9/L" "10^9/L" "10^9/L" "g/dL" ...
#>   ..$ lower_lim     : num [1:381] 1.8 1.4 0.2 11.9 0.37 150 1.8 1.4 0.2 11.9 ...
#>   ..$ upper_lim     : num [1:381] 7.2 4 1 14.5 0.47 400 7.2 4 1 14.5 ...
#>   ..$ item_value    : num [1:381] 4.16 1.7 0.12 11.5 0.25 ...
#>   ..$ significance  : Factor w/ 5 levels "within limits",..: 2 1 2 2 2 2 2 1 2 2 ...
#>   ..$ reason_notdone: chr [1:381] NA NA NA NA ...
#>   ..$ region        : chr [1:381] "BEL" "BEL" "BEL" "BEL" ...
#>   ..$ value_scaled  : num [1:381] 0.437 0.115 -0.1 -0.154 -1.2 ...
#>   ..$ out_of_lim    : Factor w/ 2 levels "1","0": 2 2 1 1 1 1 1 2 1 1 ...
#>   ..$ text_label    : chr [1:381] "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 4.16 10^9/L" "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 1.7 10^9/L" "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 0.12 10^9/L" "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 11.5 g/dL" ...
#>  $ Conc. Procedures: cnc_prcd [28 x 22] (S3: conc_procedures/common_forms/tbl_df/tbl/data.frame)
#>   ..$ site_code     : chr [1:28] "BEL07" "BEL07" "BEL07" "BEL07" ...
#>   ..$ subject_id    : chr [1:28] "BEL_07_645" "BEL_07_645" "BEL_07_645" "BEL_07_645" ...
#>   ..$ event_id      : chr [1:28] "COMMON_PR" "COMMON_PR" "COMMON_PR" "COMMON_PR" ...
#>   ..$ event_date    : Date[1:28], format: "2023-09-14" "2023-09-14" ...
#>   ..$ event_repeat  : int [1:28] 1 1 1 1 1 1 2 2 2 2 ...
#>   ..$ form_id       : chr [1:28] "CP" "CP" "CP" "CP" ...
#>   ..$ form_repeat   : int [1:28] 1 1 1 1 1 1 2 2 2 2 ...
#>   ..$ edit_date_time: POSIXct[1:28], format: "2023-09-14 09:37:00" "2023-09-14 09:37:00" ...
#>   ..$ day           : 'difftime' num [1:28] 101 101 101 101 ...
#>   .. ..- attr(*, "units")= chr "days"
#>   ..$ event_name    : chr [1:28] "Any visit" "Any visit" "Any visit" "Any visit" ...
#>   ..$ event_label   : Factor w/ 8 levels "SCR","V1","V2",..: NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ form_type     : chr [1:28] "common_forms" "common_forms" "common_forms" "common_forms" ...
#>   ..$ item_name     : chr [1:28] "CP Number" "CP Name" "CP Indication" "CP Start Date" ...
#>   ..$ item_type     : chr [1:28] "other" "other" "other" "other" ...
#>   ..$ item_group    : chr [1:28] "Conc. Procedures" "Conc. Procedures" "Conc. Procedures" "Conc. Procedures" ...
#>   ..$ item_unit     : chr [1:28] "(unit missing)" "(unit missing)" "(unit missing)" "(unit missing)" ...
#>   ..$ lower_lim     : num [1:28] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ upper_lim     : num [1:28] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ item_value    : chr [1:28] "1" "Appendectomy" "Prophylaxis" "2019-NK-NK" ...
#>   ..$ significance  : chr [1:28] NA NA NA NA ...
#>   ..$ reason_notdone: chr [1:28] NA NA NA NA ...
#>   ..$ region        : chr [1:28] "BEL" "BEL" "BEL" "BEL" ...
#>   ..- attr(*, "synch_time")= chr "2023-09-15 10:10:00 UTC"
#>  $ Electrolytes    : continus [275 x 25] (S3: continuous/tbl_df/tbl/data.frame)
#>   ..$ site_code     : chr [1:275] "BEL04" "BEL04" "BEL04" "BEL04" ...
#>   ..$ subject_id    : chr [1:275] "BEL_04_133" "BEL_04_133" "BEL_04_133" "BEL_04_133" ...
#>   ..$ event_id      : chr [1:275] "SCR" "SCR" "SCR" "SCR" ...
#>   ..$ event_date    : Date[1:275], format: "2023-06-05" "2023-06-05" ...
#>   ..$ event_repeat  : int [1:275] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ form_id       : chr [1:275] "LBSER" "LBSER" "LBSER" "LBSER" ...
#>   ..$ form_repeat   : int [1:275] 1 1 1 1 1 1 2 2 2 2 ...
#>   ..$ edit_date_time: POSIXct[1:275], format: "2023-06-23 09:00:00" "2023-06-23 09:00:00" ...
#>   ..$ day           : 'difftime' num [1:275] 0 0 0 0 ...
#>   .. ..- attr(*, "units")= chr "days"
#>   ..$ event_name    : chr [1:275] "Screening" "Screening" "Screening" "Screening" ...
#>   ..$ event_label   : Factor w/ 8 levels "SCR","V1","V2",..: 1 1 1 1 1 1 2 2 2 2 ...
#>   ..$ form_type     : chr [1:275] "study_forms" "study_forms" "study_forms" "study_forms" ...
#>   ..$ item_name     : Factor w/ 6 levels "Sodium","Potassium",..: 1 2 3 4 5 6 1 2 3 4 ...
#>   ..$ item_type     : chr [1:275] "continuous" "continuous" "continuous" "continuous" ...
#>   ..$ item_group    : chr [1:275] "Electrolytes" "Electrolytes" "Electrolytes" "Electrolytes" ...
#>   ..$ item_unit     : chr [1:275] "mmol/L" "mmol/L" "mmol/L" "mmol/L" ...
#>   ..$ lower_lim     : num [1:275] 136 3 98 22 2.15 0.66 136 3 98 22 ...
#>   ..$ upper_lim     : num [1:275] 145 4.5 107 29 2.55 1.07 145 4.5 107 29 ...
#>   ..$ item_value    : num [1:275] 138 4.5 105 20.5 2.26 0.87 138 4.14 103 23 ...
#>   ..$ significance  : Factor w/ 5 levels "within limits",..: 1 2 1 3 1 1 1 2 1 1 ...
#>   ..$ reason_notdone: chr [1:275] NA NA NA NA ...
#>   ..$ region        : chr [1:275] "BEL" "BEL" "BEL" "BEL" ...
#>   ..$ value_scaled  : num [1:275] 0.222 1 0.778 -0.214 0.275 ...
#>   ..$ out_of_lim    : Factor w/ 2 levels "0","1": 1 1 1 2 1 1 1 1 1 1 ...
#>   ..$ text_label    : chr [1:275] "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 138 mmol/L" "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 4.5 mmol/L" "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 105 mmol/L" "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 20.5 mmol/L" ...
#>  $ General         : general [261 x 22] (S3: general/tbl_df/tbl/data.frame)
#>   ..$ site_code     : chr [1:261] "BEL04" "BEL04" "BEL04" "BEL04" ...
#>   ..$ subject_id    : chr [1:261] "BEL_04_133" "BEL_04_133" "BEL_04_133" "BEL_04_133" ...
#>   ..$ event_id      : chr [1:261] "SCR" "SCR" "SCR" "SCR" ...
#>   ..$ event_date    : Date[1:261], format: "2023-06-05" "2023-06-05" ...
#>   ..$ event_repeat  : int [1:261] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ form_id       : chr [1:261] "DM" "DM" "DM" "DM" ...
#>   ..$ form_repeat   : int [1:261] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ edit_date_time: POSIXct[1:261], format: "2023-06-23 08:31:00" "2023-06-23 08:31:00" ...
#>   ..$ day           : 'difftime' num [1:261] 0 0 0 0 ...
#>   .. ..- attr(*, "units")= chr "days"
#>   ..$ event_name    : chr [1:261] "Screening" "Screening" "Screening" "Screening" ...
#>   ..$ event_label   : Factor w/ 8 levels "SCR","V1","V2",..: 1 1 1 1 1 1 1 1 1 2 ...
#>   ..$ form_type     : chr [1:261] "general" "general" "general" "general" ...
#>   ..$ item_name     : chr [1:261] "Age" "Sex" "ChildbearingPotential" "MenopauseReason" ...
#>   ..$ item_type     : chr [1:261] "other" "other" "other" "other" ...
#>   ..$ item_group    : chr [1:261] "General" "General" "General" "General" ...
#>   ..$ item_unit     : chr [1:261] "(unit missing)" "(unit missing)" "(unit missing)" "(unit missing)" ...
#>   ..$ lower_lim     : num [1:261] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ upper_lim     : num [1:261] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ item_value    : chr [1:261] "88" "Male" "No" "Postmenopausal" ...
#>   ..$ significance  : chr [1:261] NA NA NA NA ...
#>   ..$ reason_notdone: chr [1:261] NA NA NA NA ...
#>   ..$ region        : chr [1:261] "BEL" "BEL" "BEL" "BEL" ...
#>   ..- attr(*, "synch_time")= chr "2023-09-15 10:10:00 UTC"
#>  $ Liver function  : continus [180 x 25] (S3: continuous/tbl_df/tbl/data.frame)
#>   ..$ site_code     : chr [1:180] "BEL04" "BEL04" "BEL04" "BEL04" ...
#>   ..$ subject_id    : chr [1:180] "BEL_04_133" "BEL_04_133" "BEL_04_133" "BEL_04_133" ...
#>   ..$ event_id      : chr [1:180] "SCR" "SCR" "SCR" "SCR" ...
#>   ..$ event_date    : Date[1:180], format: "2023-06-05" "2023-06-05" ...
#>   ..$ event_repeat  : int [1:180] 1 1 1 1 1 1 1 1 3 3 ...
#>   ..$ form_id       : chr [1:180] "LBSER" "LBSER" "LBSER" "LBSER" ...
#>   ..$ form_repeat   : int [1:180] 1 1 1 1 2 2 2 2 3 3 ...
#>   ..$ edit_date_time: POSIXct[1:180], format: "2023-06-23 09:00:00" "2023-07-05 12:33:00" ...
#>   ..$ day           : 'difftime' num [1:180] 0 0 0 0 ...
#>   .. ..- attr(*, "units")= chr "days"
#>   ..$ event_name    : chr [1:180] "Screening" "Screening" "Screening" "Screening" ...
#>   ..$ event_label   : Factor w/ 8 levels "SCR","V1","V2",..: 1 1 1 1 2 2 2 2 4 4 ...
#>   ..$ form_type     : chr [1:180] "study_forms" "study_forms" "study_forms" "study_forms" ...
#>   ..$ item_name     : Factor w/ 4 levels "Bilirubin","AST",..: 1 2 3 4 1 2 3 4 1 2 ...
#>   ..$ item_type     : chr [1:180] "continuous" "continuous" "continuous" "continuous" ...
#>   ..$ item_group    : chr [1:180] "Liver function" "Liver function" "Liver function" "Liver function" ...
#>   ..$ item_unit     : chr [1:180] "µmol/L" "U/L" "U/L" "U/L" ...
#>   ..$ lower_lim     : num [1:180] 0 10 10 5 0 10 10 5 0 10 ...
#>   ..$ upper_lim     : num [1:180] 21 35 25 36 21 35 25 36 21 35 ...
#>   ..$ item_value    : num [1:180] 6 21 35 24 11 32 45 25 8 43 ...
#>   ..$ significance  : Factor w/ 5 levels "within limits",..: 1 2 2 1 1 1 2 1 1 2 ...
#>   ..$ reason_notdone: chr [1:180] NA NA NA NA ...
#>   ..$ region        : chr [1:180] "BEL" "BEL" "BEL" "BEL" ...
#>   ..$ value_scaled  : num [1:180] 0.286 0.44 1.667 0.613 0.524 ...
#>   ..$ out_of_lim    : Factor w/ 2 levels "0","1": 1 1 2 1 1 1 2 1 1 2 ...
#>   ..$ text_label    : chr [1:180] "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 6 µmol/L" "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 21 U/L" "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 35 U/L" "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 24 U/L" ...
#>  $ Medical History : mdcl_hst [757 x 22] (S3: medical_history/common_forms/tbl_df/tbl/data.frame)
#>   ..$ site_code     : chr [1:757] "BEL04" "BEL04" "BEL04" "BEL04" ...
#>   ..$ subject_id    : chr [1:757] "BEL_04_133" "BEL_04_133" "BEL_04_133" "BEL_04_133" ...
#>   ..$ event_id      : chr [1:757] "COMMON_MH" "COMMON_MH" "COMMON_MH" "COMMON_MH" ...
#>   ..$ event_date    : Date[1:757], format: "2023-06-23" "2023-06-23" ...
#>   ..$ event_repeat  : int [1:757] 1 1 1 1 1 1 1 1 2 2 ...
#>   ..$ form_id       : chr [1:757] "MH" "MH" "MH" "MH" ...
#>   ..$ form_repeat   : int [1:757] 1 1 1 1 1 1 1 1 2 2 ...
#>   ..$ edit_date_time: POSIXct[1:757], format: "2023-06-23 09:12:00" "2023-06-23 09:12:00" ...
#>   ..$ day           : 'difftime' num [1:757] 18 18 18 18 ...
#>   .. ..- attr(*, "units")= chr "days"
#>   ..$ event_name    : chr [1:757] "Any visit" "Any visit" "Any visit" "Any visit" ...
#>   ..$ event_label   : Factor w/ 8 levels "SCR","V1","V2",..: NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ form_type     : chr [1:757] "common_forms" "common_forms" "common_forms" "common_forms" ...
#>   ..$ item_name     : chr [1:757] "MH Number" "MH Name" "MH Ongoing" "MH Treatment" ...
#>   ..$ item_type     : chr [1:757] "other" "other" "other" "other" ...
#>   ..$ item_group    : chr [1:757] "Medical History" "Medical History" "Medical History" "Medical History" ...
#>   ..$ item_unit     : chr [1:757] "(unit missing)" "(unit missing)" "(unit missing)" "(unit missing)" ...
#>   ..$ lower_lim     : num [1:757] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ upper_lim     : num [1:757] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ item_value    : chr [1:757] "1" "Epilepsy" "Yes" "Medication" ...
#>   ..$ significance  : chr [1:757] NA NA NA NA ...
#>   ..$ reason_notdone: chr [1:757] NA NA NA NA ...
#>   ..$ region        : chr [1:757] "BEL" "BEL" "BEL" "BEL" ...
#>   ..- attr(*, "synch_time")= chr "2023-09-15 10:10:00 UTC"
#>  $ Medication      : medicatn [2,856 x 22] (S3: medication/common_forms/tbl_df/tbl/data.frame)
#>   ..$ site_code     : chr [1:2856] "BEL04" "BEL04" "BEL04" "BEL04" ...
#>   ..$ subject_id    : chr [1:2856] "BEL_04_133" "BEL_04_133" "BEL_04_133" "BEL_04_133" ...
#>   ..$ event_id      : chr [1:2856] "COMMON_CM" "COMMON_CM" "COMMON_CM" "COMMON_CM" ...
#>   ..$ event_date    : Date[1:2856], format: "2023-06-23" "2023-06-23" ...
#>   ..$ event_repeat  : int [1:2856] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ form_id       : chr [1:2856] "CM" "CM" "CM" "CM" ...
#>   ..$ form_repeat   : int [1:2856] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ edit_date_time: POSIXct[1:2856], format: "2023-06-23 09:17:00" "2023-06-23 09:17:00" ...
#>   ..$ day           : 'difftime' num [1:2856] 18 18 18 18 ...
#>   .. ..- attr(*, "units")= chr "days"
#>   ..$ event_name    : chr [1:2856] "Any visit" "Any visit" "Any visit" "Any visit" ...
#>   ..$ event_label   : Factor w/ 8 levels "SCR","V1","V2",..: NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ form_type     : chr [1:2856] "common_forms" "common_forms" "common_forms" "common_forms" ...
#>   ..$ item_name     : chr [1:2856] "CM Number" "CM Name" "CM Indication" "CM Ongoing" ...
#>   ..$ item_type     : chr [1:2856] "other" "other" "other" "other" ...
#>   ..$ item_group    : chr [1:2856] "Medication" "Medication" "Medication" "Medication" ...
#>   ..$ item_unit     : chr [1:2856] "(unit missing)" "(unit missing)" "(unit missing)" "(unit missing)" ...
#>   ..$ lower_lim     : num [1:2856] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ upper_lim     : num [1:2856] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ item_value    : chr [1:2856] "1" "Paracetamol (Gliigestatrilat)" "Prophylaxis" "Yes" ...
#>   ..$ significance  : chr [1:2856] NA NA NA NA ...
#>   ..$ reason_notdone: chr [1:2856] NA NA NA NA ...
#>   ..$ region        : chr [1:2856] "BEL" "BEL" "BEL" "BEL" ...
#>   ..- attr(*, "synch_time")= chr "2023-09-15 10:10:00 UTC"
#>  $ Renal function  : continus [135 x 25] (S3: continuous/tbl_df/tbl/data.frame)
#>   ..$ site_code     : chr [1:135] "BEL04" "BEL04" "BEL04" "BEL04" ...
#>   ..$ subject_id    : chr [1:135] "BEL_04_133" "BEL_04_133" "BEL_04_133" "BEL_04_133" ...
#>   ..$ event_id      : chr [1:135] "SCR" "SCR" "SCR" "VIS1" ...
#>   ..$ event_date    : Date[1:135], format: "2023-06-05" "2023-06-05" ...
#>   ..$ event_repeat  : int [1:135] 1 1 1 1 1 1 3 3 3 5 ...
#>   ..$ form_id       : chr [1:135] "LBSER" "LBSER" "LBSER" "LBSER" ...
#>   ..$ form_repeat   : int [1:135] 1 1 1 2 2 2 3 3 3 4 ...
#>   ..$ edit_date_time: POSIXct[1:135], format: "2023-08-18 15:39:00" "2023-08-18 15:46:00" ...
#>   ..$ day           : 'difftime' num [1:135] 0 0 0 30 ...
#>   .. ..- attr(*, "units")= chr "days"
#>   ..$ event_name    : chr [1:135] "Screening" "Screening" "Screening" "Visit 1" ...
#>   ..$ event_label   : Factor w/ 8 levels "SCR","V1","V2",..: 1 1 1 2 2 2 4 4 4 6 ...
#>   ..$ form_type     : chr [1:135] "study_forms" "study_forms" "study_forms" "study_forms" ...
#>   ..$ item_name     : Factor w/ 3 levels "Creatinine","eGFR",..: 1 2 3 1 2 3 1 2 3 1 ...
#>   ..$ item_type     : chr [1:135] "continuous" "continuous" "continuous" "continuous" ...
#>   ..$ item_group    : chr [1:135] "Renal function" "Renal function" "Renal function" "Renal function" ...
#>   ..$ item_unit     : chr [1:135] "µmol/L" "ml/min/1.73 m²" "µmol/L" "µmol/L" ...
#>   ..$ lower_lim     : num [1:135] 45 60 140 45 60 2.9 45 60 2.9 45 ...
#>   ..$ upper_lim     : num [1:135] 84 90 340 84 90 8.2 84 90 8.2 84 ...
#>   ..$ item_value    : num [1:135] 58 48 288 110 89 4.1 61 90 7.5 117 ...
#>   ..$ significance  : Factor w/ 5 levels "within limits",..: 2 2 1 2 2 1 2 2 1 2 ...
#>   ..$ reason_notdone: chr [1:135] NA NA NA NA ...
#>   ..$ region        : chr [1:135] "BEL" "BEL" "BEL" "BEL" ...
#>   ..$ value_scaled  : num [1:135] 0.333 -0.4 0.74 1.667 0.967 ...
#>   ..$ out_of_lim    : Factor w/ 2 levels "0","1": 1 2 1 2 1 1 1 1 1 2 ...
#>   ..$ text_label    : chr [1:135] "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 58 µmol/L" "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 48 ml/min/1.73 m²" "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 288 µmol/L" "<b>BEL_04_133</b>\\n2023-07-05\\nVisit 1 (day 30)\\nValue: 110 µmol/L" ...
#>  $ Response        : tibble [6 x 22] (S3: tbl_df/tbl/data.frame)
#>   ..$ site_code     : chr [1:6] "BEL04" "BEL04" "BEL04" "BEL04" ...
#>   ..$ subject_id    : chr [1:6] "BEL_04_133" "BEL_04_133" "BEL_04_772" "BEL_04_772" ...
#>   ..$ event_id      : chr [1:6] "VIS5" "VIS5" "VIS5" "VIS5" ...
#>   ..$ event_date    : Date[1:6], format: "2023-08-30" "2023-08-30" ...
#>   ..$ event_repeat  : int [1:6] 5 5 5 5 5 5
#>   ..$ form_id       : chr [1:6] "RS" "RS" "RS" "RS" ...
#>   ..$ form_repeat   : int [1:6] 1 1 1 1 1 1
#>   ..$ edit_date_time: POSIXct[1:6], format: "2023-09-13 15:23:00" "2023-09-13 15:23:00" ...
#>   ..$ day           : 'difftime' num [1:6] 86 86 86 86 ...
#>   .. ..- attr(*, "units")= chr "days"
#>   ..$ event_name    : chr [1:6] "Visit 5" "Visit 5" "Visit 5" "Visit 5" ...
#>   ..$ event_label   : Factor w/ 8 levels "SCR","V1","V2",..: 6 6 6 6 6 6
#>   ..$ form_type     : chr [1:6] "study_forms" "study_forms" "study_forms" "study_forms" ...
#>   ..$ item_name     : chr [1:6] "Response" "Responder" "Response" "Responder" ...
#>   ..$ item_type     : chr [1:6] "other" "other" "other" "other" ...
#>   ..$ item_group    : chr [1:6] "Response" "Response" "Response" "Response" ...
#>   ..$ item_unit     : chr [1:6] "(unit missing)" "(unit missing)" "(unit missing)" "(unit missing)" ...
#>   ..$ lower_lim     : num [1:6] NA NA NA NA NA NA
#>   ..$ upper_lim     : num [1:6] NA NA NA NA NA NA
#>   ..$ item_value    : chr [1:6] "Stable disease (SD)" "No" "Cured" "Yes" ...
#>   ..$ significance  : chr [1:6] NA NA NA NA ...
#>   ..$ reason_notdone: chr [1:6] NA NA NA NA ...
#>   ..$ region        : chr [1:6] "BEL" "BEL" "BEL" "BEL" ...
#>   ..- attr(*, "synch_time")= chr "2023-09-15 10:10:00 UTC"
#>  $ Vital signs     : continus [478 x 25] (S3: continuous/tbl_df/tbl/data.frame)
#>   ..$ site_code     : chr [1:478] "BEL04" "BEL04" "BEL04" "BEL04" ...
#>   ..$ subject_id    : chr [1:478] "BEL_04_133" "BEL_04_133" "BEL_04_133" "BEL_04_133" ...
#>   ..$ event_id      : chr [1:478] "SCR" "SCR" "SCR" "SCR" ...
#>   ..$ event_date    : Date[1:478], format: "2023-06-05" "2023-06-05" ...
#>   ..$ event_repeat  : int [1:478] 1 1 1 1 1 1 1 1 1 1 ...
#>   ..$ form_id       : chr [1:478] "VS" "VS" "VS" "VS" ...
#>   ..$ form_repeat   : int [1:478] 1 1 1 1 1 1 1 2 2 2 ...
#>   ..$ edit_date_time: POSIXct[1:478], format: "2023-06-23 08:32:00" "2023-06-23 08:32:00" ...
#>   ..$ day           : 'difftime' num [1:478] 0 0 0 0 ...
#>   .. ..- attr(*, "units")= chr "days"
#>   ..$ event_name    : chr [1:478] "Screening" "Screening" "Screening" "Screening" ...
#>   ..$ event_label   : Factor w/ 8 levels "SCR","V1","V2",..: 1 1 1 1 1 1 1 2 2 2 ...
#>   ..$ form_type     : chr [1:478] "study_forms" "study_forms" "study_forms" "study_forms" ...
#>   ..$ item_name     : Factor w/ 8 levels "Systolic blood pressure",..: 8 7 3 4 5 1 2 8 7 3 ...
#>   ..$ item_type     : chr [1:478] "continuous" "continuous" "continuous" "continuous" ...
#>   ..$ item_group    : chr [1:478] "Vital signs" "Vital signs" "Vital signs" "Vital signs" ...
#>   ..$ item_unit     : chr [1:478] "kg" "kg/m2" "beats/min" "breaths/min" ...
#>   ..$ lower_lim     : num [1:478] 45 18.5 60 12 35 90 55 45 18.5 60 ...
#>   ..$ upper_lim     : num [1:478] 200 30 100 20 38.5 160 90 200 30 100 ...
#>   ..$ item_value    : num [1:478] 50 24.5 60 19 36.7 ...
#>   ..$ significance  : Factor w/ 5 levels "within limits",..: 1 1 1 1 1 1 1 1 3 1 ...
#>   ..$ reason_notdone: chr [1:478] NA NA NA NA ...
#>   ..$ region        : chr [1:478] "BEL" "BEL" "BEL" "BEL" ...
#>   ..$ value_scaled  : num [1:478] 0.0323 0.5235 0 0.875 0.4857 ...
#>   ..$ out_of_lim    : Factor w/ 2 levels "0","1": 1 1 1 1 1 1 1 1 2 1 ...
#>   ..$ text_label    : chr [1:478] "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 50 kg" "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 24.52 kg/m2" "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 60 beats/min" "<b>BEL_04_133</b>\\n2023-06-05\\nScreening (day 0)\\nValue: 19 breaths/min" ...
}\if{html}{\out{</div>}}
}
}
\source{
See \code{raw-data/clinsightful_data.R} for details.
}
\usage{
cs_app_data
}
\description{
A list of data frames split by \code{item_group}. It is currently
not needed to provide the application data in this format. However, the
function can be useful for clinsight data customization, especially in
combination with \code{\link[=create_table]{create_table()}}. It creates table classes of the
\code{item_group} name (in lower case letters, and all punctuation replaced by an
underscore), as long as there is a method available with the corresponding
item_group name for the function \code{\link[=create_table]{create_table()}}. If all data is of type
'continuous', then laboratory values such as lab limits and units are
cleaned.
}
\keyword{datasets}
