% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs_available_data}
\alias{cs_available_data}
\title{Application 'Available Data'}
\format{
a data.frame with 1974 rows and
6 variables.

\if{html}{\out{<div class="sourceCode r">}}\preformatted{str(cs_available_data)
#> tibble [1,974 x 6] (S3: tbl_df/tbl/data.frame)
#>  $ subject_id : chr [1:1974] "BEL_04_772" "BEL_04_772" "BEL_04_772" "BEL_04_772" ...
#>  $ item_name  : chr [1:1974] "Hypotension" "Atrial Fibrillation (N: 2)" "Tachycardia" "Urinary Tract Infection" ...
#>  $ form_repeat: int [1:1974] 1 2 3 4 5 1 1 1 3 2 ...
#>  $ item_group : chr [1:1974] "Adverse events" "Adverse events" "Adverse events" "Adverse events" ...
#>  $ event_name : chr [1:1974] "Any visit" "Any visit" "Any visit" "Any visit" ...
#>  $ event_label: chr [1:1974] "Any visit" "Any visit" "Any visit" "Any visit" ...
}\if{html}{\out{</div>}}
}
\source{
See \code{raw-data/clinsightful_data.R} for details.
}
\usage{
cs_available_data
}
\description{
Data frame containing info about available data per individual,
such as visits, adverse events, etc. Will be used in module
\code{\link[=mod_queries_server]{mod_queries_server()}}, to select available items to create a query for per
individual and per form. Required columns are the ones distinctively
identifying an item. For now that are site_code, event_name, subject_id,
event_label, item_group, item_name.
}
\keyword{datasets}
