% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_helpers.R
\name{datatable_custom}
\alias{datatable_custom}
\title{Custom interactive datatable}
\usage{
datatable_custom(
  data,
  rename_vars = NULL,
  title = NULL,
  selection = "single",
  extensions = c("Scroller", "ColReorder"),
  dom = "fti",
  options = list(),
  allow_listing_download = NULL,
  export_label = NULL,
  escape = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame to use within the application.}

\item{rename_vars}{An optional named character vector. If provided, it will
rename any column names found in this vector to the provided name.}

\item{title}{Optional. Character string with the title of the table.}

\item{selection}{See \code{\link[DT:datatable]{DT::datatable()}}. Default set to 'single'.}

\item{extensions}{See \code{\link[DT:datatable]{DT::datatable()}}. Default set to 'Scroller'.}

\item{dom}{See \url{https://datatables.net/reference/option/dom}. A div
element will be inserted before the table for the table title. Default set
to 'fti' resulting in 'f<"header h5">ti'.}

\item{options}{See \code{\link[DT:datatable]{DT::datatable()}}. Must be a list.
\itemize{
\item Modifiable defaults:
\itemize{
\item \code{scrollY = '400px'}
\item \code{scrollX = TRUE}
\item \code{scroller = TRUE}
\item \code{deferREnder = TRUE}
\item \code{scrollResize = TRUE}
\item \code{scrollCollapse = TRUE}
\item \code{colReorder = TRUE}
}
\item Non-modifiable defaults:
\itemize{
\item \code{dom}: Defined by the \code{dom} parameter.
\item \code{initComplete}: Defaults to a function to insert table title into dataTable container.
}
}}

\item{allow_listing_download}{Logical, whether to allow the user to download
the table as an Excel file. Defaults to the \code{allow_listing_download}
configuration option in \code{golem-config.yml}, but can be overwritten here if
needed.}

\item{export_label}{Character string with the table export label. Only used
for downloadable tables (if \code{allow_listing_download} is \code{TRUE}).}

\item{escape}{Whether to escape HTML entities in the table. See \code{\link[DT:datatable]{DT::datatable()}}.}

\item{...}{Other optional arguments that will be passed to \code{\link[DT:datatable]{DT::datatable()}}.}
}
\value{
A \code{DT::datatable} object.
}
\description{
Small wrapper around \code{\link[DT:datatable]{DT::datatable()}}. Will be used to create tables in a
consistent style.
}
\examples{
\dontrun{
datatable_custom(mtcars)
}
}
\keyword{internal}
