% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_helpers.R
\name{datatable_custom}
\alias{datatable_custom}
\title{Custom interactive datatable}
\usage{
datatable_custom(
  data,
  rename_vars = NULL,
  title = NULL,
  selection = "single",
  ...
)
}
\arguments{
\item{data}{A data frame to use within the application.}

\item{rename_vars}{An optional named character vector. If provided, it will
rename any column names found in this vector to the provided name.}

\item{title}{Optional. Character string with the title of the table.}

\item{selection}{See \code{\link[DT:datatable]{DT::datatable()}}. Default set to 'single'.}

\item{...}{Other optional arguments that will be parsed to \code{\link[DT:datatable]{DT::datatable()}}.}
}
\value{
A \code{DT::datatable} object.
}
\description{
Small wrapper around \code{\link[DT:datatable]{DT::datatable()}}. Will be used to create tables in a
consistent style.
}
\examples{
datatable_custom(mtcars)
}
