% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_add_log}
\alias{db_add_log}
\title{Add Logging Table}
\usage{
db_add_log(con)
}
\arguments{
\item{con}{A DBI Connection to the SQLite DB}
}
\description{
Both creates the logging table and the trigger to update it for
all_review_data.
}
\keyword{internal}
