% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_add_log}
\alias{db_add_log}
\title{Add Logging Table}
\usage{
db_add_log(con, keys = c("id", idx_cols))
}
\arguments{
\item{con}{A DBI Connection to the SQLite DB}

\item{keys}{A character vector specifying which columns should not be updated
in a table. Defaults to 'id' and the package-defined index columns
(\code{idx_cols}).}
}
\description{
Both creates the logging table and the trigger to update it for
all_review_data.
}
\keyword{internal}
