% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_create}
\alias{db_create}
\title{Create app database}
\usage{
db_create(data, db_path, reviewed = "No", reviewer = "", status = "new")
}
\arguments{
\item{data}{Either a data frame with review data (Usually created with
\code{\link[=get_review_data]{get_review_data()}}), or a character path to the raw data files.}

\item{db_path}{A character vector with the path to the database to be
created.}

\item{reviewed}{Character vector. Sets the reviewed tag in the review
database.}

\item{reviewer}{Character vector. Sets the reviewer in the review database.
Defaults to \code{Admin}.}

\item{status}{Character vector. Sets the status in the review database.
Defaults to \code{new}.}
}
\value{
A database will be created. Nothing else will be returned.
}
\description{
Creates application database. To create a database with all data flagged as
'new', use the default settings of \code{reviewed}, \code{reviewer}, and \code{status}.
}
\seealso{
\code{\link[=get_review_data]{get_review_data()}}
}
