% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_get_review}
\alias{db_get_review}
\title{Retrieve review}
\usage{
db_get_review(db_path, ..., db_table = "all_review_data")
}
\arguments{
\item{db_path}{Character vector. Needs to be a valid path to a database.}

\item{...}{Named arguments specifying which records to retrieve, see
examples. Note that \code{...} will be processed with \code{data.frame()} and thus
the arguments within \code{...} should be convertible to a data frame. This is
chosen so that filters of length one can be used with other filters since
they will be recycled (for example, when selecting multiple events of one
subject).}

\item{db_table}{Character string. Name of the table to collect. Will only be
used if \code{data} is a character string to a database.}
}
\value{
A data frame.
}
\description{
Small helper function to retrieve the (latest) review data from the database
with the given subject id (\code{subject}) and \code{form}.
}
\keyword{internal}
