% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_get_summary_data}
\alias{db_get_summary_data}
\title{Get summary data}
\usage{
db_get_summary_data(db_path, forms_to_review, subjects_to_review)
}
\arguments{
\item{db_path}{Character vector. Needs to be a valid path to a database.}

\item{forms_to_review}{A character vector with all forms that need to be
reviewed.}

\item{subjects_to_review}{A character vector with the subject ids that should
be selected.}
}
\value{
A data frame with summary data
}
\description{
Get summary data
}
\keyword{internal}
