% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_save_review}
\alias{db_save_review}
\title{Save review in database}
\usage{
db_save_review(rv_records, db_path, table = "all_review_data")
}
\arguments{
\item{rv_records}{A data frame containing the rows of data that needs to be
checked.}

\item{db_path}{Character vector. Path to the database.}

\item{table}{Character vector. Names of the table within the database to
save the review in.}
}
\value{
Review information will be written in the database. No local objects
will be returned.
}
\description{
Helper function to save review in database. All old data will not be changed.
New rows with the new/updated review data will be added to the applicable
database tables.
}
