% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_update}
\alias{db_update}
\title{Update app database}
\usage{
db_update(
  data,
  db_path,
  common_vars = c("subject_id", "event_name", "item_group", "form_repeat", "item_name"),
  edit_time_var = "edit_date_time",
  query_path = NULL
)
}
\arguments{
\item{data}{An updated data frame with review data.}

\item{db_path}{Character vector. Path to the database.}

\item{common_vars}{A character vector containing the common key variables.}

\item{edit_time_var}{A character vector with the column name of the edit-time
variable.}

\item{query_path}{If it exists, provide a file.path to an RDS file,
containing data.frame with the same schema as \code{query_data_skeleton}, used
to populate a new database with existing queries from the EDC.}
}
\value{
Nothing will be returned.
}
\description{
Compares the latest edit date-times in the review database and in the data
frame. If the provided data frame is newer, the database will be updated.
}
