% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_update}
\alias{db_update}
\title{Update app database}
\usage{
db_update(
  data,
  db_path,
  key_cols = key_columns,
  edit_time_var = "edit_date_time"
)
}
\arguments{
\item{data}{An updated data frame with review data.}

\item{db_path}{Character vector. Path to the database.}

\item{key_cols}{A character vector containing the common key variables.
Defaults to \code{ClinSight} \code{\link[=key_columns]{key_columns()}}.}

\item{edit_time_var}{A character vector with the column name of the edit-time
variable.}
}
\value{
Nothing will be returned.
}
\description{
Compares the latest edit date-times in the review database and in the data
frame. If the provided data frame is newer, the database will be updated.
}
