% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_update}
\alias{db_update}
\title{Update app database}
\usage{
db_update(
  data,
  db_path,
  common_vars = c("subject_id", "event_name", "item_group", "form_repeat", "item_name"),
  edit_time_var = "edit_date_time",
  data_synched = FALSE
)
}
\arguments{
\item{data}{An updated data frame with review data.}

\item{db_path}{Character vector. Path to the database.}

\item{common_vars}{A character vector containing the common key variables.}

\item{edit_time_var}{A character vector with the column name of the edit-time
variable.}

\item{data_synched}{Logical. Whether the database was synched or not. If
TRUE, the synchronization date stored in the database will be updated to
the current day.}
}
\value{
Nothing will be returned.
}
\description{
Compares the latest edit date-times in the review database and in the data
frame. If the provided data frame is newer, the database will be updated.
}
