% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_upsert}
\alias{db_upsert}
\title{UPSERT to all_review_data}
\usage{
db_upsert(con, data, common_cols)
}
\arguments{
\item{con}{A DBI Connection to the SQLite DB}

\item{data}{A data frame containing the data to UPSERT into all_review_data}

\item{common_cols}{A character vector specifying which columns define a
unique row}
}
\description{
UPSERT to all_review_data
}
\keyword{internal}
