% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_SQLite.R
\name{db_upsert}
\alias{db_upsert}
\title{UPSERT to all_review_data}
\usage{
db_upsert(con, data, idx_cols)
}
\arguments{
\item{con}{A DBI Connection to the SQLite DB}

\item{data}{A data frame containing the data to UPSERT into all_review_data}

\item{idx_cols}{A character vector specifying which columns define a
unique index for a row}
}
\value{
invisibly returns TRUE. Is run for it's side effects on the DB.
}
\description{
Performs an UPSERT on all_review_data. New records will be appended to the
table. Changed/updated records will be applied to the table based on the
index column constraint.
}
\keyword{internal}
