% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_figures.R
\name{fig_timeline}
\alias{fig_timeline}
\title{Create timeline}
\usage{
fig_timeline(data, meta_data = metadata)
}
\arguments{
\item{data}{Data frame to use.}

\item{meta_data}{Metadata file with all possible time points.}
}
\value{
A ggplot2 object.
}
\description{
Function to create a simple timeline figure using \code{ggplot2}.
}
