% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_figures.R
\name{fig_timeline}
\alias{fig_timeline}
\title{Create timeline}
\usage{
fig_timeline(data, events)
}
\arguments{
\item{data}{Data frame to use.}

\item{events}{Data frame containing information about all events. Used
to create the right labels in the timeline figure.}
}
\value{
A ggplot2 object.
}
\description{
Function to create a simple timeline figure using \code{ggplot2}.
}
\keyword{internal}
