% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_figures.R
\name{fig_timeseries}
\alias{fig_timeseries}
\title{time series figure}
\usage{
fig_timeseries(
  data,
  xval = "day",
  id = "subject_id",
  id_to_highlight,
  color_fill = "significance",
  point_size = "reviewed",
  label = "text_label",
  show_all_participants = TRUE,
  show_all_hover_labels = FALSE,
  scale = FALSE,
  use_unscaled_limits = FALSE
)
}
\arguments{
\item{data}{A data frame containing the data of interest.}

\item{xval}{x-value to be used.}

\item{id}{Character string. Column containing the unique Patient IDs.}

\item{id_to_highlight}{Character vector. Unique Patient ID. Connects the
lines in the time series.}

\item{color_fill}{Character vector. Column name of the variable to be used
for filling color in the figure.}

\item{point_size}{character vector. Column in the data frame that controls
the point size in the figure.}

\item{label}{Character vector. Label to be used for each data point. Will
only be visible if the ggplot object is converted to an interactive plot
using \code{plotly::ggplotly()}. See \code{\link[=plotly_figure]{plotly_figure()}}}

\item{show_all_participants}{Logical to toggle background patterns.}

\item{show_all_hover_labels}{Logical to toggle hover labels.}

\item{scale}{A logical. Whether to us a scaled value (value_scaled) or the
raw variable (item_value).}

\item{use_unscaled_limits}{If TRUE, limits provided in the data frame will be
used. This parameter will be ignored if scaled is set to \code{TRUE}.}
}
\value{
A faceted ggplot2 time series figure.
}
\description{
Creates a ggplot2 faceted time series line plot. Can be called within
\code{\link[=plotly_figure]{plotly_figure()}} to create an interactive figure.
}
\examples{
\dontrun{
set.seed(2025)
mock_data <- lapply(paste0("Subject", 1:10) , \(x){
  data.frame(
    subject_id = x,
    day = sample(1:25, 10),
    item_name = sample(c("item1", "item2"), 10, replace = TRUE),
    item_value = runif(10, 0 , 50),
    significance = sample(
      c("limits unknown", "out of limits, clinically significant", 
        "out of limits, clinically insignificant"), 
      10, 
      replace = TRUE
    ),
    text_label = "test text",
    reviewed = sample(c("Yes", "No"), replace = TRUE)
  )
}) |>
  dplyr::bind_rows()
fig_timeseries(mock_data, id_to_highlight = "Subject10")
}
}
\seealso{
\code{\link[=plotly_figure]{plotly_figure()}}
}
\keyword{internal}
