% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_helpers.R
\name{fix_multiple_choice_vars}
\alias{fix_multiple_choice_vars}
\title{Correct multiple choice variables}
\usage{
fix_multiple_choice_vars(
  data = raw_data,
  meta = metadata,
  var_column = "var",
  value_column = "item_value",
  suffix = "[[:digit:]]+$",
  common_vars = c("subject_id", "event_repeat", "event_date", "form_repeat"),
  collapse_with = "; "
)
}
\arguments{
\item{data}{data frame (typically the raw data)}

\item{meta}{metadata, list of data frames.}

\item{var_column}{column name in which the variable names are stored}

\item{value_column}{column name in which the values of the variables are stored}

\item{suffix}{Multiple choice suffix. Used to define multiple choice values}

\item{common_vars}{variables used for identifying unique rows in the dataset.}

\item{collapse_with}{character value to collapse the multiple choice options with.
If this value is NULL, the rows will be left as is.}
}
\value{
data frame with corrected multiple choice variables
}
\description{
Function to correct multiple choice variables in the data.
}
