% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_appdata.R
\name{get_appdata}
\alias{get_appdata}
\title{Get appdata

Converts data to a list of data frames in the expected format to be used by the
the Shiny application.}
\usage{
get_appdata(data, meta = metadata)
}
\arguments{
\item{data}{A data frame, commonly raw data merged with metadata.}

\item{meta}{A data frame with metadata. Used to convert the data set to the
appropriate format.}
}
\value{
A list of data frames.
}
\description{
Get appdata

Converts data to a list of data frames in the expected format to be used by the
the Shiny application.
}
