% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_form_tbl_fct_helpers.R
\name{get_form_table}
\alias{get_form_table}
\title{Get form table}
\usage{
get_form_table(
  form_data,
  form_review_data,
  form,
  form_items,
  transformation = "none",
  value_column = "item_value",
  unit_column = "item_unit",
  active_subject,
  pending_form_review_status = NULL,
  is_SAE = NULL,
  key_cols = key_columns
)
}
\arguments{
\item{form_data}{A data frame with the study data of the respective form.}

\item{form_review_data}{A data frame with he review data of the form.}

\item{form}{A character string with the name of the form.}

\item{form_items}{Named character vector with all form_items to display.}

\item{active_subject}{A character string with the active subject id.}

\item{pending_form_review_status}{A logical, indicating whether all items of
the entire form for the active subject id are reviewed. Note that the
logical can have three statuses, namely TRUE (=reviewed), FALSE (=not
reviewed), and indeterminate (=partially reviewed).}

\item{is_SAE}{A logical, indicating whether the form is a SAE form. If TRUE,
will make some adjustments to the columns to display.}

\item{key_cols}{Columns that identify a unique row in the data. Defaults to
\code{ClinSight} \code{\link[=key_columns]{key_columns()}}.}
}
\description{
Get form table
}
\keyword{internal}
