% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_form_tbl_fct_helpers.R
\name{get_form_table}
\alias{get_form_table}
\title{Get form table}
\usage{
get_form_table(
  form_data,
  form_review_data,
  form,
  form_items,
  active_subject,
  updated,
  is_SAE,
  id_cols = idx_cols
)
}
\arguments{
\item{form_data}{}

\item{form_review_data}{}

\item{form}{}

\item{active_subject}{}

\item{updated}{}

\item{is_SAE}{}

\item{id_cols}{}
}
\description{
Get form table
}
\keyword{internal}
