% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_form_tbl_fct_helpers.R
\name{get_form_table}
\alias{get_form_table}
\title{Get form table}
\usage{
get_form_table(
  form_data,
  form_review_data,
  form,
  form_items,
  active_subject,
  is_reviewed = NULL,
  is_SAE = NULL,
  id_cols = idx_cols
)
}
\arguments{
\item{form_data}{A data frame with the study data of the respective form.}

\item{form_review_data}{A data frame with he review data of the form.}

\item{form}{A character string with the name of the form.}

\item{form_items}{Named character vector with all form_items to display.}

\item{active_subject}{A character string with the active subject id.}

\item{is_reviewed}{A logical, indicating whether all items of the entire form
for the active subject id are reviewed.}

\item{is_SAE}{A logical, indicating whether the form is a SAE form. If TRUE,
will make some adjustments to the columns to display.}

\item{id_cols}{Columns that identify a unique row in the data.}
}
\description{
Get form table
}
\keyword{internal}
