% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_appdata.R
\name{get_raw_csv_data}
\alias{get_raw_csv_data}
\title{Get raw data from CSV files}
\usage{
get_raw_csv_data(
  data_path = Sys.getenv("RAW_DATA_PATH"),
  exclude = c("README.csv$", "Pending_forms.csv$", "MEDRA.csv$", "WHODrug.csv$",
    "_Queries.csv$"),
  delim = ",",
  skip = 1
)
}
\arguments{
\item{data_path}{Path to the folder that contains the CSV files with the
application data.}

\item{exclude}{character vector with regex that identify csv files that
should not be merged in study data. For example, readme files included in
the export.}

\item{delim}{Delimiter to use to read in files.}

\item{skip}{Number of rows to skip when reading in files.}
}
\value{
A data frame with raw application data.
}
\description{
Get raw data from CSV files
}
