% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_data_fct_helpers.R
\name{get_review_data}
\alias{get_review_data}
\title{Get review data}
\usage{
get_review_data(
  data,
  common_vars = c("subject_id", "event_name", "item_group", "form_repeat", "item_name"),
  date_var = "event_date",
  edit_time_var = "edit_date_time"
)
}
\arguments{
\item{data}{A data frame, commonly raw data merged with meta data.}

\item{common_vars}{A character vector with the common variable column names.
Assumes that these variables can identify a unique data point that needs to
be reviewed.}

\item{date_var}{A character vector with the name of the date column.}

\item{edit_time_var}{A character vector with the name of the edit date-time
column.}
}
\value{
A data frame with review data.
}
\description{
Create a data frame containing all data to be reviewed. Selects the \emph{latest}
edit date-time for every event/item.
}
\keyword{internal}
