% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_appdata_summary_tables.R
\name{get_timeline_data}
\alias{get_timeline_data}
\title{Get timeline data}
\usage{
get_timeline_data(
  data,
  table_data,
  timeline_cols = c("subject_id", "event_name", "form_repeat", "item_group", "start",
    "group", "end", "title", "className", "id", "order"),
  treatment_label = "💊 Tₓ"
)
}
\arguments{
\item{data}{A list of data frames, with compatible clinical trial data.}

\item{table_data}{A list of data frames containing clinical trial data in
wide format. Created with \code{\link[=create_table]{create_table()}}.}

\item{timeline_cols}{Character vector with the name of the columns of the
output data frame.}

\item{treatment_label}{Character vector with the label to use for the
treatment item in the timeline.}
}
\value{
A data frame with timeline data.
}
\description{
Function to create timeline data, to be used in a \code{timevis::timeline()}
object.
}
