% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_merge_item_pairs.R
\name{merge_item_pair}
\alias{merge_item_pair}
\title{Merge a pair of items}
\usage{
merge_item_pair(
  data,
  item_name,
  item_name_other,
  merge_action = c("combine", "replace"),
  name_column = "item_name",
  value_column = "item_value",
  id_cols = idx_cols
)
}
\arguments{
\item{data}{A data frame.}

\item{item_name}{A character string with the name of the item to merge.}

\item{item_name_other}{A character string with the name of the item to merge with.}

\item{merge_action}{A character string with the action to take when merging.}

\item{name_column}{A character string with the name of the column containing the item names.}

\item{value_column}{A character string with the name of the column containing the item values.}

\item{id_cols}{A character vector with the names of the columns that uniquely identify a row.}
}
\value{
A data frame with the merged items.
}
\description{
Merge a pair of items
}
\keyword{internal}
