% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Application Metadata}
\format{
\subsection{\code{metadata}: A list with 14 objects.}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{str(metadata)
#> List of 14
#>  $ column_names    : tibble [11 x 2] (S3: tbl_df/tbl/data.frame)
#>   ..$ name_raw: chr [1:11] "SiteCode" "SubjectId" "EventId" "EventDate" ...
#>   ..$ name_new: chr [1:11] "site_code" "subject_id" "event_id" "event_date" ...
#>  $ events          : tibble [38 x 10] (S3: tbl_df/tbl/data.frame)
#>   ..$ event_id               : chr [1:38] "SCR" "VIS1" "VIS2" "VIS3" ...
#>   ..$ event_id_pattern       : chr [1:38] "^SCR$" "^VIS1$" "^VIS2$" "^VIS3$" ...
#>   ..$ is_regular_visit       : logi [1:38] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>   ..$ event_label_custom     : chr [1:38] NA "C1D1" "C1D13" "C2D1" ...
#>   ..$ event_name_custom      : chr [1:38] "Screening" "Visit 1" "Visit 2" "Visit 3" ...
#>   ..$ is_baseline_event      : logi [1:38] TRUE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ generate_labels        : logi [1:38] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ meta_event_order       : int [1:38] 1 2 3 4 5 6 7 8 9 10 ...
#>   ..$ add_visit_number       : logi [1:38] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ add_event_repeat_number: logi [1:38] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ events_other    : tibble [9 x 6] (S3: tbl_df/tbl/data.frame)
#>   ..$ event_id        : chr [1:9] NA "C1D1" "C1D13" "C2D1" ...
#>   ..$ event_id_pattern: chr [1:9] "^Screening" NA NA NA ...
#>   ..$ expected_visits : chr [1:9] "1" "1" "1" "1" ...
#>   ..$ event_name      : chr [1:9] "Screening" "C1D1" "C1D13" "C2D1" ...
#>   ..$ event_name_short: chr [1:9] NA NA NA NA ...
#>   ..$ add_event_number: chr [1:9] "FALSE" "FALSE" "FALSE" "FALSE" ...
#>  $ events_old      : tibble [19 x 3] (S3: tbl_df/tbl/data.frame)
#>   ..$ event_number: chr [1:19] "0" "1" "2" "3" ...
#>   ..$ event_name  : chr [1:19] "Screening" "Visit 1" "Visit 2" "Visit 3" ...
#>   ..$ event_label : chr [1:19] "V0" "V1" "V2" "V3" ...
#>  $ common_forms    :'data.frame':	57 obs. of  4 variables:
#>   ..$ item_name : chr [1:57] "AE Number" "AE Name" "AESI" "AE start date" ...
#>   ..$ item_type : chr [1:57] "other" "other" "other" "other" ...
#>   ..$ item_group: chr [1:57] "Adverse events" "Adverse events" "Adverse events" "Adverse events" ...
#>   ..$ merge_with: chr [1:57] NA NA NA NA ...
#>  $ study_forms     :'data.frame':	93 obs. of  6 variables:
#>   ..$ item_name  : chr [1:93] "Tranf. Date" "Type" "Type other" "Number of units" ...
#>   ..$ item_type  : chr [1:93] "other" "other" "other" "other" ...
#>   ..$ item_group : chr [1:93] "Transfusions" "Transfusions" "Transfusions" "Transfusions" ...
#>   ..$ unit       : chr [1:93] NA NA NA NA ...
#>   ..$ lower_limit: chr [1:93] NA NA NA NA ...
#>   ..$ upper_limit: chr [1:93] NA NA NA NA ...
#>  $ general         :'data.frame':	39 obs. of  3 variables:
#>   ..$ item_name : chr [1:39] "Age" "Sex" "ECOG" "Eligible" ...
#>   ..$ item_type : chr [1:39] "other" "other" "other" "other" ...
#>   ..$ item_group: chr [1:39] "General" "General" "General" "General" ...
#>  $ form_level_data :'data.frame':	16 obs. of  4 variables:
#>   ..$ item_group         : chr [1:16] "Adverse events" "Medication" "Conc. Procedures" "Medical History" ...
#>   ..$ item_scale         : logi [1:16] NA NA NA NA NA NA ...
#>   ..$ use_unscaled_limits: logi [1:16] NA NA NA NA NA NA ...
#>   ..$ review_required    : logi [1:16] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  $ table_names     : tibble [15 x 2] (S3: tbl_df/tbl/data.frame)
#>   ..$ table_name: chr [1:15] "Edit date" "Date" "Event" "Event" ...
#>   ..$ raw_name  : chr [1:15] "edit_date_time" "event_date" "event_label" "event_name" ...
#>  $ settings        :List of 4
#>   ..$ pre_pivot_fns  : chr "apply_study_specific_suffix_fixes"
#>   ..$ post_pivot_fns : chr "apply_edc_specific_changes"
#>   ..$ post_merge_fns : chr "apply_study_specific_fixes"
#>   ..$ treatment_label: chr "💊"
#>  $ conversion_table: tibble [103 x 7] (S3: tbl_df/tbl/data.frame)
#>   ..$ var_name              : chr [1:103] "Neutrophils" "Neutrophils" "Neutrophils" "Neutrophils" ...
#>   ..$ unit_code             : chr [1:103] "1" "2" "3" "4" ...
#>   ..$ unit                  : chr [1:103] "10^9/L" "G/L" "Gpt/L" "cells/L" ...
#>   ..$ units_used            : chr [1:103] NA NA NA NA ...
#>   ..$ unit_conversion       : chr [1:103] "1" "1" "1" "1/(10^9)" ...
#>   ..$ unit_standard         : chr [1:103] "10^9/L" "10^9/L" "10^9/L" "10^9/L" ...
#>   ..$ unit_conversion_source: chr [1:103] "https://unitslab.com/node/75" "https://unitslab.com/node/75" "https://unitslab.com/node/75" "https://unitslab.com/node/75" ...
#>  $ Sheet2          : tibble [138 x 2] (S3: tbl_df/tbl/data.frame)
#>   ..$ item_name: chr [1:138] "ALT" "ALT" "ALT" "AST" ...
#>   ..$ item_unit: chr [1:138] "µkat/L" "U/L" "IU/L" "µkat/L" ...
#>  $ Sheet1          : tibble [18 x 4] (S3: tbl_df/tbl/data.frame)
#>   ..$ Unit   : chr [1:18] "10^9/L" "G/L" "Gpt/L" "10^3/µl" ...
#>   ..$ ...2   : chr [1:18] NA NA NA NA ...
#>   ..$ ...3   : chr [1:18] "1" "1" "1" "1" ...
#>   ..$ Synonym: chr [1:18] "10^9/L" "10^9/L" "10^9/L" "10^9/L" ...
#>  $ items_expanded  : tibble [503 x 9] (S3: tbl_df/tbl/data.frame)
#>   ..$ form_type  : chr [1:503] "common_forms" "common_forms" "common_forms" "common_forms" ...
#>   ..$ var        : chr [1:503] "AE_AESPID" "AE_AETERM" "AE_AESI" "AE_AESTDAT" ...
#>   ..$ suffix     : chr [1:503] NA NA NA NA ...
#>   ..$ item_name  : chr [1:503] "AE Number" "AE Name" "AESI" "AE start date" ...
#>   ..$ item_type  : chr [1:503] "other" "other" "other" "other" ...
#>   ..$ item_group : chr [1:503] "Adverse events" "Adverse events" "Adverse events" "Adverse events" ...
#>   ..$ unit       : chr [1:503] NA NA NA NA ...
#>   ..$ lower_limit: chr [1:503] NA NA NA NA ...
#>   ..$ upper_limit: chr [1:503] NA NA NA NA ...
}\if{html}{\out{</div>}}
}
}
\source{
Can be created with an Excel file. The Excel file format is chosen so
that the metadata can be changed easily per study. See
\code{raw-data/metadata.R} for details.
}
\usage{
metadata
}
\description{
A list of data frames and settings containing metadata that will be used for
the application. The metadata will be merged with with raw data. It controls
the variables that will be used in the application, and in which tab the
variables will be shown. The goal is that most, if not all, study-specific
data will be captured in the metadata, leaving the scripts to run the
application largely unaltered between studies.See \code{vignette("Metadata")} for
in-depth information.
}
\keyword{datasets}
