% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Application Metadata}
\format{
\subsection{\code{metadata}: A list with 9 objects.}{

\if{html}{\out{<div class="sourceCode r">}}\preformatted{str(metadata)
#> List of 9
#>  $ column_names   : tibble [10 x 2] (S3: tbl_df/tbl/data.frame)
#>   ..$ name_raw: chr [1:10] "SiteCode" "SubjectId" "EventId" "EventDate" ...
#>   ..$ name_new: chr [1:10] "site_code" "subject_id" "event_id" "event_date" ...
#>  $ events         : tibble [10 x 10] (S3: tbl_df/tbl/data.frame)
#>   ..$ event_id               : chr [1:10] "SCR" "START" "AF" "CFUEPRO_6M" ...
#>   ..$ event_id_pattern       : chr [1:10] "^SCR$" "^START$" "^AF$" "^CFUEPRO_6M$" ...
#>   ..$ is_regular_visit       : logi [1:10] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>   ..$ event_label_custom     : chr [1:10] NA "Start" "AF Abl." "CFU 6M" ...
#>   ..$ event_name_custom      : chr [1:10] "Screening" "Study Start" "Early AF Ablation" "CFU 6M" ...
#>   ..$ is_baseline_event      : logi [1:10] FALSE TRUE FALSE FALSE FALSE FALSE ...
#>   ..$ generate_labels        : logi [1:10] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ meta_event_order       : int [1:10] 1 2 3 4 5 6 7 8 9 10
#>   ..$ add_visit_number       : logi [1:10] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ add_event_repeat_number: logi [1:10] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ common_forms   :'data.frame':	98 obs. of  4 variables:
#>   ..$ item_name : chr [1:98] "AE Number" "AE Name" "AESI" "AE start date" ...
#>   ..$ item_type : chr [1:98] "other" "other" "other" "other" ...
#>   ..$ item_group: chr [1:98] "Adverse events" "Adverse events" "Adverse events" "Adverse events" ...
#>   ..$ merge_with: chr [1:98] NA NA NA NA ...
#>  $ study_forms    :'data.frame':	44 obs. of  7 variables:
#>   ..$ item_name  : chr [1:44] "Date planned ablation" "Not done" "Hospital admission Date" "Hospital discharge Date" ...
#>   ..$ item_type  : chr [1:44] "other" "other" "other" "other" ...
#>   ..$ item_group : chr [1:44] "Ablation" "Ablation" "Ablation" "Ablation" ...
#>   ..$ merge_with : chr [1:44] NA "ABLA_ND_SPEC" NA NA ...
#>   ..$ unit       : chr [1:44] NA NA NA NA ...
#>   ..$ lower_limit: chr [1:44] NA NA NA NA ...
#>   ..$ upper_limit: chr [1:44] NA NA NA NA ...
#>  $ general        :'data.frame':	24 obs. of  3 variables:
#>   ..$ item_name : chr [1:24] "Age" "Sex" "ECOG" "Eligible" ...
#>   ..$ item_type : chr [1:24] "other" "other" "other" "other" ...
#>   ..$ item_group: chr [1:24] "General" "General" "General" "General" ...
#>  $ form_level_data:'data.frame':	16 obs. of  4 variables:
#>   ..$ item_group         : chr [1:16] "Adverse events" "Medication" "Demographics" "Conc. Procedures" ...
#>   ..$ item_scale         : logi [1:16] NA NA NA NA NA NA ...
#>   ..$ use_unscaled_limits: logi [1:16] NA NA NA NA NA NA ...
#>   ..$ review_required    : logi [1:16] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>  $ table_names    : tibble [16 x 2] (S3: tbl_df/tbl/data.frame)
#>   ..$ table_name: chr [1:16] "Edit date" "Date" "Event" "Event" ...
#>   ..$ raw_name  : chr [1:16] "edit_date_time" "event_date" "event_label" "event_name" ...
#>  $ settings       :List of 5
#>   ..$ pre_pivot_fns  : chr "apply_study_specific_suffix_fixes"
#>   ..$ post_pivot_fns : chr "apply_edc_specific_changes"
#>   ..$ post_merge_fns : chr "apply_study_specific_fixes"
#>   ..$ treatment_label: chr "💊 Tₓ"
#>   ..$ study_name     : chr "EASThigh-AFNET 11"
#>  $ items_expanded : tibble [204 x 9] (S3: tbl_df/tbl/data.frame)
#>   ..$ form_type  : chr [1:204] "common_forms" "common_forms" "common_forms" "common_forms" ...
#>   ..$ var        : chr [1:204] "AE_AESPID" "AE_AETERM" "AE_AESI" "AE_AESTDAT" ...
#>   ..$ suffix     : chr [1:204] NA NA NA NA ...
#>   ..$ item_name  : chr [1:204] "AE Number" "AE Name" "AESI" "AE start date" ...
#>   ..$ item_type  : chr [1:204] "other" "other" "other" "other" ...
#>   ..$ item_group : chr [1:204] "Adverse events" "Adverse events" "Adverse events" "Adverse events" ...
#>   ..$ unit       : chr [1:204] NA NA NA NA ...
#>   ..$ lower_limit: chr [1:204] NA NA NA NA ...
#>   ..$ upper_limit: chr [1:204] NA NA NA NA ...
}\if{html}{\out{</div>}}
}
}
\source{
Can be created with an Excel file. The Excel file format is chosen so
that the metadata can be changed easily per study. See
\code{raw-data/metadata.R} for details.
}
\usage{
metadata
}
\description{
A list of data frames and settings containing metadata that will be used for
the application. The metadata will be merged with with raw data. It controls
the variables that will be used in the application, and in which tab the
variables will be shown. The goal is that most, if not all, study-specific
data will be captured in the metadata, leaving the scripts to run the
application largely unaltered between studies.See \code{vignette("Metadata")} for
in-depth information.
}
\keyword{datasets}
