% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metadata}
\alias{metadata}
\title{Application Metadata}
\format{
\subsection{\code{metadata} A list with the following data frames:}{
\itemize{
\item \code{events}: Contains the columns \code{event_number}, \code{event_name}, and
\code{event_label}. Used to create a simple timeline in the application, with
predefined number of planned visits.
\item \code{common_forms}: Contains the columns \code{var}, \code{suffix}, \code{item_name} ,
\code{item_type}, \code{item_group}. Used to select and rename the variables of
interest in the common forms.
\item \code{study_forms}: Contains the same columns as the data frame \code{common_forms},
and in addition the columns \code{unit}, \code{lower_limit}, \code{uppter_limit}. Used to
select and rename the variables of interest in the common forms.
\item \code{general}: Contains the same columns as \code{common_forms}.
\item \code{groups}: Contains the columns \code{item_group}, \code{item_type}, \code{item_scale},
\code{use_unscaled_limits}.
}
}
}
\source{
Can be created with an Excel file. The Excel file format is chosen so
that the metadata can be changed easily per study. See
\code{raw-data/metadata.R} for details.
}
\usage{
metadata
}
\description{
A list of data frames containing metadata that will be used for the
application. The metadata will be merged with with raw data. It controls the
variables that will be used in the application, and in which tab the
variables will be shown. The goal is that most, if not all, study-specific
data will be captured in the metadata, leaving the scripts to run the
application largely unaltered between studies.
}
\keyword{datasets}
