% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_db_synch_info.R
\name{mod_db_synch_info_server}
\alias{mod_db_synch_info_server}
\title{Database Synchronization information - Shiny module Server}
\usage{
mod_db_synch_info_server(
  id,
  app_data,
  db_path,
  current_date = Sys.Date(),
  show_synch_warning = golem::app_prod()
)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module
Server.}

\item{app_data}{List of data frames. Contains the application data, with data
of each form stored in a data frame. Required to derive the latest EDC
date.}

\item{db_path}{Character vector with the path to the app database. Required
to retrieve the latest database synch date that is stored in the data frame
"db_synch_time".}

\item{current_date}{Current date. Standard \code{Sys.Date()}. Can be useful to set
for testing purposes.}

\item{show_synch_warning}{Logical. Whether to show a pop-up message with a
warning if database synchronization did not happen on the current day. Will
normally be shown if \code{golem::app_prod()} returns \code{TRUE}.}
}
\description{
A shiny module. Used to display database synchronization information/
}
\details{
The module is currently initiated in the main sidebar. It shows the time at
which the database was synchronized last, and the time of the latest EDC data
entry (or data edit). In addition, a warning message will pop up if the
synchronization date is either unknown or older than one day.
}
\seealso{
\code{\link[=mod_db_synch_info_ui]{mod_db_synch_info_ui()}}
}
