% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_sidebar.R
\name{mod_main_sidebar_server}
\alias{mod_main_sidebar_server}
\title{App sidebar - Shiny module Server}
\usage{
mod_main_sidebar_server(
  id,
  r,
  navinfo,
  app_data,
  app_tables,
  app_vars,
  db_path,
  forms_to_review,
  available_data
)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module
Server.}

\item{r}{Common reactivevalues.  Will be passed on to
\code{\link[=mod_query_add_server]{mod_query_add_server()}}.}

\item{navinfo}{Reactive values created with \code{\link[shiny:reactiveValues]{shiny::reactiveValues()}}. Used
to send back information about the page change to the server, when clicking
on the adverse event box. Required for \code{\link[=mod_navigate_forms_server]{mod_navigate_forms_server()}}.
Furthermore, it contains the currently active form (\code{navinfo$active_form}),
needed for \code{\link[=mod_review_forms_server]{mod_review_forms_server()}} and \code{\link[=mod_query_add_server]{mod_query_add_server()}}, and
the currently active tab (\code{navinfo$active_tab}), needed for
\code{\link[=mod_review_forms_server]{mod_review_forms_server()}}. These are reactive values containing a
character vector with the currently active tab or form.
\code{navinfo$active_tab} is also used to only show review controls when a the
tab is either common events or study data. Other active tab values are, for
example, \code{Start} (start page), or \code{Queries}.}

\item{app_data}{List of data frames. Contains the application data, with data
of each form stored in a data frame. Required to set the review
configuration in \code{\link[=mod_review_config_server]{mod_review_config_server()}}.}

\item{app_tables}{List of data frames with the app data in wide table format.
Required to set the review configuration in \code{\link[=mod_review_config_server]{mod_review_config_server()}}}

\item{app_vars}{A list with common variables found in the data and metadata.
Required to set the review configuration in \code{\link[=mod_review_config_server]{mod_review_config_server()}}.}

\item{db_path}{Character vector with the path to the app database. Required
in \code{\link[=mod_query_add_server]{mod_query_add_server()}}.}

\item{forms_to_review}{A reactive value containing a character vector with
the forms that need review of the active participant.}

\item{available_data}{A data frame containing all available data, usually
created with the function \code{\link[=get_available_data]{get_available_data()}}. The data frame will be
passed on to the module \code{\link[=mod_query_add_server]{mod_query_add_server()}}, which requires this data
frame and is embedded in \code{mod_main_sidebar_server()}.}
}
\description{
A shiny module. Used to create the sidebar in the main application.
}
\details{
This sidebar contains the main controls for the application. In the sidebar,
a panel is available which allows the user to apply a review to the active
form and patient (forms can be marked as reviewed, a comment can be added,
and the review can be saved. In addition, it contains buttons to navigate to
the next/previous forms and to add a query to the active form.
}
\seealso{
\code{\link[=mod_main_sidebar_ui]{mod_main_sidebar_ui()}}, \code{\link[=mod_query_add_server]{mod_query_add_server()}}
}
