% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_form_tbl.R
\name{mod_review_form_tbl_server}
\alias{mod_review_form_tbl_server}
\title{Review forms table - Shiny module Server}
\usage{
mod_review_form_tbl_server(
  id,
  form,
  form_data,
  form_review_data,
  form_items,
  active_subject,
  show_all,
  enable_text_wrap = reactive(FALSE),
  table_names = NULL,
  title = NULL
)
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module
Server.}

\item{form}{A character string with the name of the form to display.}

\item{form_data}{Common reactive value. Used to manage the server data
displayed in the DataTable.}

\item{form_review_data}{Common reactive value containing the review data of
the form.}

\item{form_items}{Named character vector with all form_items to display.}

\item{active_subject}{Reactive value containing the active subject id.}

\item{show_all}{Common reactive value, a logical indicating whether all
records should be displayed.}

\item{enable_text_wrap}{A reactive value, to enable/disable multi-line
table rows. Usually disabled so that deferred rendering is possible, but
can be enabled for better viewing experience.}

\item{table_names}{An optional character vector. If provided, will be used
within \code{\link[=datatable_custom]{datatable_custom()}}, to improve the column names in the final
interactive tables.}

\item{title}{An optional character vector. If provided, will be used within
\code{\link[=datatable_custom]{datatable_custom()}}, as the title for the table.}
}
\description{
Shiny module. Used to handle the logic and presentation of form review data.
}
\details{
This module handles the logic associated with the form review tables. It
handles all the logic associated with the table check boxes and the observers
needed to keep the browser table in-synch with the server table.
}
\seealso{
\code{\link[=mod_review_form_tbl_ui]{mod_review_form_tbl_ui()}}, \code{\link[=mod_common_forms_ui]{mod_common_forms_ui()}},
\code{\link[=mod_common_forms_server]{mod_common_forms_server()}}, \code{\link[=mod_study_forms_ui]{mod_study_forms_ui()}},
\code{\link[=mod_study_forms_server]{mod_study_forms_server()}}
}
