% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_timeline.R
\name{mod_timeline_server}
\alias{mod_timeline_server}
\title{Interactive timeline - Shiny module Server}
\usage{
mod_timeline_server(id, r, form, treatment_label = "💊 Tₓ")
}
\arguments{
\item{id}{Character string, used to connect the module UI with the module
Server.}

\item{r}{Common reactive values. Used to access the data frames
\code{review_data}, \code{filtered_tables}, and the active \code{subject_id}.
\code{review_data} will be used to minimize data points on the timeline that are
already reviewed.}

\item{form}{A character vector, the form in which the timeline needs to be
embedded. Currently, only the form 'Adverse events' is supported.}

\item{treatment_label}{Character with the treatment label to use. Defaults to
"💊 Tₓ" if missing.}
}
\description{
A shiny module. Creates an interactive timeline. Used to provide a quick
overview of adverse events, severe adverse events, and study activities such
as Investigational Product administration. Helpful to judge whether for
example an event is related to an Investigational Product.
}
\seealso{
\code{\link[=mod_timeline_ui]{mod_timeline_ui()}}, \code{\link[=mod_common_forms_ui]{mod_common_forms_ui()}},
\code{\link[=mod_common_forms_server]{mod_common_forms_server()}}
}
