% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data_helpers.R
\name{rename_raw_data}
\alias{rename_raw_data}
\title{Rename raw data}
\usage{
rename_raw_data(data, column_specs = metadata$column_specs)
}
\arguments{
\item{data}{A data frame with raw study data.}

\item{column_specs}{A data frame with column specifications. Should have at
least the columns \code{name_raw}, containing the current column names, and
\code{name_new}, containing the new column names. \code{name_new} should contain all
names that are required for ClinSight to function properly
(\code{required_col_names}).}
}
\value{
A data frame
}
\description{
Helper function to rename raw data
}
