% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(
  onStart = NULL,
  options = list(),
  enableBookmarking = NULL,
  uiPattern = "/",
  meta = metadata,
  data = clinsightful_data,
  user_db = "user_db.sqlite",
  credentials_db = "credentials_db.sqlite",
  credentials_pwd = Sys.getenv("DB_SECRET"),
  test_mode = FALSE,
  ...
)
}
\arguments{
\item{onStart}{A function that will be called before the app is actually run.
This is only needed for \code{shinyAppObj}, since in the \code{shinyAppDir}
case, a \code{global.R} file can be used for this purpose.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}

\item{enableBookmarking}{Can be one of \code{"url"}, \code{"server"}, or
\code{"disable"}. The default value, \code{NULL}, will respect the setting from
any previous calls to  \code{\link[shiny:enableBookmarking]{enableBookmarking()}}. See \code{\link[shiny:enableBookmarking]{enableBookmarking()}}
for more information on bookmarking your app.}

\item{uiPattern}{A regular expression that will be applied to each \code{GET}
request to determine whether the \code{ui} should be used to handle the
request. Note that the entire request path must match the regular
expression in order for the match to be considered successful.}

\item{meta}{A data frame containing metadata.}

\item{data}{Either a data frame or a character string with the path to the
app data in .rds format.}

\item{user_db}{Character string. Path to the app database. If not existing,
will be created based on app data and metadata, with all data labeled as
'new'/not yet reviewed.}

\item{credentials_db}{Character string. Path to the credentials database.}

\item{credentials_pwd}{Character string with the credentials' database
password.}

\item{test_mode}{Logical, whether to run the application in test mode.}

\item{...}{arguments to pass to golem_opts. See \code{?golem::get_golem_options}
for more details.}
}
\description{
Run the Shiny Application
}
