% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_test_clinsight.R
\name{test_clinsight}
\alias{test_clinsight}
\title{Test ClinSight}
\usage{
test_clinsight(
  clinsight_data = clinsightful_data,
  meta_data = metadata,
  clinsight_config = "test"
)
}
\arguments{
\item{clinsight_data}{A data frame with (ClinSight compatible) study data. If
not provided, the example study data in the ClinSight package will be used.}

\item{meta_data}{A list of data frames with (ClinSight compatible) metadata.
If not provided, the example metadata in the ClinSight package will be used}

\item{clinsight_config}{A character vector with the ClinSight configuration
to use.}
}
\value{
Will start the ClinSight Shiny application
}
\description{
Provides a wrapper around \code{\link[=run_app]{run_app()}}, to test ClinSight with different
configurations. Unlike the function \code{\link[=run_app]{run_app()}}, this function stores any
data in a temporary folder, keeping the environment clean.
}
\details{
While this function is useful for testing and development purposes, for use
in production the function \code{\link[=run_app]{run_app()}} should be used directly.
}
