% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_review_data_fct_helpers.R
\name{update_review_data}
\alias{update_review_data}
\title{Update review data}
\usage{
update_review_data(
  review_df,
  latest_review_data,
  key_cols = NULL,
  edit_time_var = "edit_date_time",
  update_time = time_stamp()
)
}
\arguments{
\item{review_df}{Data frame containing old review data that need to be
updated.}

\item{latest_review_data}{Data frame containing latest review data.}

\item{key_cols}{A character vector containing the common key variables. Defaults
to \link{key_columns} if unset.}

\item{edit_time_var}{A character vector with the column name of the edit-time
variable.}

\item{update_time}{Time stamp given to rows with new rows that are added.
Defaults to \code{\link[=time_stamp]{time_stamp()}}.}
}
\value{
A data frame containing only the rows with updated review data.
}
\description{
Update review data
}
\keyword{internal}
